% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.alignment.columns.R
\name{align.alignment.columns}
\alias{align.alignment.columns}
\title{Left Justify Alignment}
\usage{
align.alignment.columns(phylip.alignment, npaces = 6)
}
\arguments{
\item{phylip.alignment}{phylip.alignment = filename or path to sequence alignment in sequential phylip format}

\item{nspaces}{Minimum number of spaces separating sequences from sequence names (default is 6), i.e., the number of spaces between the longest sequence name and the first alignment site.}
}
\description{
This function is used for asthetics only. Uses the linux commands "column" and "sed" to ensure that columns of the input alignment are lined up (left justified) in the text file
even if sequence names are of different lengths. NOTE: This function doesn't quite work yet because there is a limit on the character length of columns, and DNA seqs are often too long.
}
