% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeStatsTable.R
\name{align.shared.loci}
\alias{align.shared.loci}
\title{Align Shared Loci

Takes as input a set of fasta sequence files, one per species (or individual). Perfoms multiple sequence alignment (using MAFFT) for loci that are shared among all species (determined by sequence names).
Returns a list of DNAStringSet objects, each of which holds aligned sequences for a locus. Optionally writes alignments in fasta format (one alignment per file).}
\usage{
align.shared.loci(
  input.seqs,
  indv,
  reference.indv = 1,
  seqname.str.delim = "_",
  seqname.str.loc = 3,
  output.dir = NULL,
  i.start = 1,
  i.stop = NA
)
}
\arguments{
\item{input.seqs}{Character vector of filepaths to the fasta sequence files. Each input file holds sequence data for one species (or for one individual if multiple individuals per species). Fasta headers (sequence names) must have a consistent, underscore-delimited format, with locus name being one of the entries (Default is for locus name to be held between the second underscore and either the end of the header or another underscore). Example header that I used for UCEs: "Genus_species_uce-XXXX", where XXXX is a number unique to a particular UCE homolog.}

\item{indv}{Character vector of names of individuals (or species if one individual per species) associated with input sequences. Order of names in indv must correspond to order of files in input.seqs, and each name in indv needs to occur in its corresponding filename in input.seqs.}

\item{reference.indv}{Number indicating which individual in indv is the reference (aka primary) individual for sequence alignment. Default is 1.}

\item{seqname.str.delim}{Character delimiter to use for parsing fasta headers of input.seqs. Default is "_". After parsing headers by this delimiter, one of the entries (set by the seqname.str.loc parameter) holds the locus name of the sequence.}

\item{seqname.str.loc}{A number specifying which part of fasta sequence header lines (after parsing by seqname.str.delim) contains the locus name. Default is 3. For example: ">Something_Anything_Locus1" or "Text_MoreText_Locus1_MoreInfo" would both be appropriate with default seqname.str.loc value of 3.}

\item{output.dir}{Where to save the output fasta alignment files. Default is NULL.}

\item{i.start}{First locus to start at. Default is 1. For now only works if set to 1.}

\item{i.stop}{Last locus to include. Default is NA, meaning to include all loci from i.start.}
}
\value{
Separate DNA alignments for each locus in the input stats table, a concatenated-locus alignment, and a partition file for the concatenated locus alignment.
}
\description{
Align Shared Loci

Takes as input a set of fasta sequence files, one per species (or individual). Perfoms multiple sequence alignment (using MAFFT) for loci that are shared among all species (determined by sequence names).
Returns a list of DNAStringSet objects, each of which holds aligned sequences for a locus. Optionally writes alignments in fasta format (one alignment per file).
}
