% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeStatsTable.R
\name{alignment.stats}
\alias{alignment.stats}
\title{Get Alignment Stats Table

Similar to the function makeExomeStatsTable, except this function takes as input the set of exon alignments rather than the unaligned exomes. The output is an abbreviated version of the stats table produce by the function makeExomeStatsTable.
This function is superceded by makeStatsTable have been abandoned. Note that pis.new is used here rather than the pis function that was used in makeExomeStatsTable function.}
\usage{
alignment.stats(align.dir, outfile, seqs.format = "phylip")
}
\arguments{
\item{align.dir}{Directory where exon alignments are located.}

\item{outfile}{Path where the output stats table should be saved.}

\item{seqs.format}{File format of input exon alignments.}
}
\value{
Abbreviated exome stats table writen to a file. The stats table is also returned as a character matrix.
}
\description{
Get Alignment Stats Table

Similar to the function makeExomeStatsTable, except this function takes as input the set of exon alignments rather than the unaligned exomes. The output is an abbreviated version of the stats table produce by the function makeExomeStatsTable.
This function is superceded by makeStatsTable have been abandoned. Note that pis.new is used here rather than the pis function that was used in makeExomeStatsTable function.
}
