% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bma.R
\name{amb}
\alias{amb}
\title{Expansion of IUPAC nucleotide symbols}
\usage{
amb(
  base,
  forceToLower = TRUE,
  checkBase = TRUE,
  IUPAC = REEs::s2v("acgturymkswbdhvn"),
  u2t = TRUE
)
}
\arguments{
\item{base}{An IUPAC symbol for a nucleotide as a single character}

\item{forceToLower}{If TRUE (the default) the base is forced to lower case.}

\item{checkBase}{If TRUE (the default) the character is checked to belong to the allowed IUPAC symbol list.}

\item{IUPAC}{A character vector containing the allowed IUPAC symbols.}

\item{u2t}{If TRUE (the default), "u" for uracil in RNA are changed into "t" for thymine in DNA.}
}
\value{
When base is missing, the list of IUPAC symbols is returned, otherwise a vector with expanded symbols.
}
\description{
This function returns the list of nucleotide matching a given IUPAC nucleotide symbol, for instance c("c", "g") for "s".
The function is a slightly modified version (internal syntax only) of the function amb in the package seqinr.
}
