% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrowLine.R
\name{arrowLine}
\alias{arrowLine}
\title{Plot Arrow}
\usage{
arrowLine(x0, y0, x1, y1 = y0, nArrow = 1, ...)
}
\arguments{
\item{x0}{x start position}

\item{y0}{y start position}

\item{x1}{x end position}

\item{y1}{y end position}

\item{nArrow}{arrow length}
}
\value{

}
\description{
Draw a line with arrows in it; e.g. -->-->-->
This was from a comment by Stack Overflow user Sacha Epskamp
}
