% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick.loci.R
\name{auto.flanking.lengths}
\alias{auto.flanking.lengths}
\title{Calculate Flanking Lengths

Internal function for the pick.loci function; implemented when flanking.length = "auto". Calculates lengths of flanking regions to target for a targetted REE, given the length of the REE, bait length, and bait tiling.
Function is applied to each REE being considered in pick.loci.}
\usage{
auto.flanking.lengths(locus.len, bt.len, bt.tiling)
}
\arguments{
\item{locus.length}{Numerical vector of loci lengths. Must be length one if type ="both"}

\item{target.length}{Length of region that includes the locus and is covered by baits.}
}
\value{
List with two numerical vectors holding lengths of upstream and downstream regions targetted.
}
\description{
Calculate Flanking Lengths

Internal function for the pick.loci function; implemented when flanking.length = "auto". Calculates lengths of flanking regions to target for a targetted REE, given the length of the REE, bait length, and bait tiling.
Function is applied to each REE being considered in pick.loci.
}
