% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick.loci.R
\name{bait.coverage.length}
\alias{bait.coverage.length}
\title{Calculate Bait Coverage Length and Number of Baits

Internal function used in pick.loci function.}
\usage{
bait.coverage.length(locus.len, bt.len, bt.tiling)
}
\arguments{
\item{locus.len}{Inherited from values in pick.loci function. This is the length of the ith REE locus.}

\item{bt.len}{Inherited from bait.length argument of pick.loci function. This is the nucleotide length of baits.}

\item{bt.tiling}{Inherited from bait.tiling argument of pick.loci function. This is the fraction of overlap between adjacent baits; value of 0.5 corresponds to 50\% overlap between adjacent baits.}
}
\value{
Named numeric with length of region covered by the minimum number of adjacent baits needed to cover locus.len; names() is equal to the number of baits.
}
\description{
Calculate Bait Coverage Length and Number of Baits

Internal function used in pick.loci function.
}
