% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bait.coverage.R
\name{bait.coverage}
\alias{bait.coverage}
\title{Bait Coverage}
\usage{
bait.coverage(
  targets,
  baits,
  distance = 0,
  min.depth = 1,
  bait.length = 120,
  fraction = T
)
}
\arguments{
\item{targets}{Set of target sequences.}

\item{baits}{Bait sequences. Multiple baits can be associated with one target sequence, but each bait is specific to one target.}

\item{distance}{Linear distance (number of nucleotides) allowed between target sites and bait sequence sites (i.e., interval on contig between the target and bait) for a target site to be considered "covered" by the bait. Default is 0.}

\item{min.depth}{This argument is not currently implemented. Future versions of this code might allow for calculating bait coverage at a user-specified minimum depth.}

\item{bait.length}{Nucleotide length of bait sequences.}

\item{fraction}{Report coverage as a fraction of the target sequence covered for each target sequence. Default is TRUE. If set to FALSE, the absolute number of covered target sites is returned for each input target sequence.}
}
\value{
A vector containing coverage (fraction or number of sites) for each input target sequence.
}
\description{
For an input set of sequences and baits targetting those sequences, this function returns either the fraction of sites or number of sites with at least 1X bait coverage for each input target sequence.
}
