% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.install.R
\name{blast.install}
\alias{blast.install}
\title{Install NCBI BLAST+}
\usage{
blast.install(install.loc = "auto", source = F)
}
\arguments{
\item{install.loc}{Path to directory where source files should be downloaded and blast should be installed. Default is to install blast in a directory called called blast-mafft, which is in the REEs package directory.}

\item{source}{Whether or not blast should be built and installed from source. Default is FALSE, in which case the function installs from either linux, windows, or macOS tarball depending on the result of Sys.info()}
}
\value{
NULL. BLAST is installed to install.loc be installed to
}
\description{
This function installs BLAST to one of several default directories or to user-specified directory.
}
