% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{blast}
\alias{blast}
\title{Run blast from R

Wrapper for running BLAST from R
Requires a local copy of BLAST, which can be dowloaded from here: https://ftp.ncbi.nlm.nih.gov/blast/executables/blast+/LATEST/
If blasting against genomes, this function is best implemented on a cluster/cloud environment, rather than locally, unless you have a super-computer...}
\usage{
blast(
  blast.path = "auto",
  method,
  subject,
  query,
  table.out = NULL,
  eval = 1e-05,
  output.format = 6,
  max.targets.per.query = 10,
  max.matches.per.target = 10,
  parallel.groups = NULL,
  num.threads = "max",
  max.runtime = 2880,
  other.args = NULL
)
}
\arguments{
\item{blast.path}{Either "auto" (the default), or the full path to the directory containing blast executables (usually the bin subfolder of a folder named "ncbi-blast-#.##.#+", where # are numbers indicating the version).
"auto" indicates that blast was installed to the REEs package folder using the function blast.install(). Future versions of this function will seach for executables on your system path, but this is not currently implemented.}

\item{method}{One of the following: "blastn", "blastp", "blastx", "deltablast", "psiblast", "rpsblast", "rpstblastn", "tblastn", "tblastx"}

\item{subject}{One of the following:
Character string indicating path to a local copy of a genome.
Character string indicating URL to fasta-formatted sequences.
XStringSet object (DNAStringSet or AAStringSet). If a XStringSet object is used, then a temporary fasta file is created to hold the sequences to search within.
If DNAStringSet or URL character string, a local copy will be saved as a temporary file.}

\item{query}{Path to the fasta file containing the query sequences to find matches for. If a XStringSet object is used, then a temporary fasta file is created to hold the query sequences.}

\item{table.out}{Where to save the output table. If NULL, the output is written to a temporary file that is read into R.}

\item{eval}{Number to use for the expect value (e-value); default is 1e-5.}

\item{output.format}{Integer indicating which format to use for the output table of matches (default 6).}

\item{max.targets.per.query}{Integer indicating the maximum number of targets (subject contigs) containing a match per query sequence. This sets the BLAST's max_target_seqs argument. Importantly, this argument does not control the total number of matches per query, because multiple matches may occer on the same subject contig.}

\item{max.matches.per.target}{Integer indicating the maximum number of matches per target (subject contig) per query. This sets BLAST's max_hsps argument. To determine the maximum total number of matches per query, multiply max.targets.per.query by max.matches.per.target.}

\item{parallel.groups}{Number of groups to split the query sequences into for running in parallel. This is experimental.}

\item{num.threads}{Either an integer indicating how many threads to use, or "max" (default), in which case num.threads is set to the number of cores available.}

\item{max.runtime}{Maximum number of minutes to run blast before quitting. Default is 2,880 (48 hours). You may need to increase this.}

\item{other.args}{A character string of the form "-argument1 value1 -argument2 value2" with additional arguments and values to pass to BLAST. See BLAST manual for definitions of available arguments. Default is NULL.}
}
\value{
Table of matches.
}
\description{
Run blast from R

Wrapper for running BLAST from R
Requires a local copy of BLAST, which can be dowloaded from here: https://ftp.ncbi.nlm.nih.gov/blast/executables/blast+/LATEST/
If blasting against genomes, this function is best implemented on a cluster/cloud environment, rather than locally, unless you have a super-computer...
}
