% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildPath.dir.R
\name{BuildPath.dir}
\alias{BuildPath.dir}
\title{Build Path of Directory}
\usage{
BuildPath.dir(path)
}
\arguments{
\item{path}{Character string with path to a directory}
}
\value{
Character vector containing paths to each directory on the input path.
}
\description{
This function returns the full path to each directory along the input path. The input path does not need to exist, but if it does, it must lead to a directory, not a file.
}
