% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDS.from.gb.R
\name{CDS.from.gb}
\alias{CDS.from.gb}
\title{Get CDS Contained in Range}
\usage{
CDS.from.gb(
  gb.object,
  targetTable.object,
  output.filename = NA,
  report.noCDS = TRUE
)
}
\description{
Obtains CDS portion of a sequence from a genbank flatfile.
Also obtains AA sequences (target and full sequence).
For some loci, the start codon is poorly annotated in ncbi, but this function determines the start codon by comparing the translated CDS to the NCBI AA sequence; makes lists of the target loci with unusual or absent annotations.
}
\details{
@param  gb.object An object of class gbRecord
@param  targetTable.object Input table containing target loci coordinates
@param  output.filename Output filename
@param  report.noCDS Logical indicating if the output object should also include two sets of amino acid sequences and information about missing/short/duplicate annotations. Default TRUE. If FALSE, the output object only includes the CDS regions contained within target loci.
@return An object of class DNAStringSet if report.noCDS=FALSE, otherwise an object of class list (length = 10) that holds one DNAStringSet object, two AAStringSet objects, and seven character vectors.
When report.noCDS = TRUE, the result is a list containing the following:
A DNAStringSet object containing CDS sequences contained within (i.e. overlapping) the regions specified in targetTable.object,
An AAStringSet object containing AA sequences coded by the CDS sequences in the DNAStringSet,
An AAStringSet object containing the full protein sequences for proteins represented in the DNAStringSet,
A vector of names of target loci for which another target has the same genomic coordinates,
A vector of names of target loci that do not overlap with sequences in the Genbank flatfile,
A vector of names of target loci without CDS annotation(s) in the Genbank record,
A vector of names of target loci for which Genbank record includes a very short CDS annotation (< 12 bp). These short CDS and AA sequences were not written to file,
A vector of names of target loci for which multiple translation frames of target CDS are found in full AA sequence,
A vector of names of target loci for which no translation frames of target CDS are found in full AA sequence,
A vector of short names for the 10 elements of the output: "CDS StringSet","Target AA StringSet","Full AA StringSet","duplicate target names","targets not in Genbank flatfile","targets without CDS annotation","targets with short CDS annotation","multiple translation frames","no matching translation frames","result names".
@export
}
