% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charTest.aa.R
\name{charTest.aa}
\alias{charTest.aa}
\title{Any AAs Present}
\usage{
charTest.aa(ls)
}
\arguments{
\item{ls}{Input amino acid sequence held in character vector}
}
\value{
TRUE or FALSE
}
\description{
Check that at least one amino acid is present in a putative amino acid sequence
}
