% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charTest.dna.R
\name{charTest.dna}
\alias{charTest.dna}
\title{Check if Nucleotides Present}
\usage{
charTest.dna(ls, method = c("any", "all"))
}
\arguments{
\item{ls}{Input DNA sequence held in character vector}

\item{method}{Either "any" or "all". If "any", then the function checks if any of the four nucleotides are present. If "all", then the function checks if all of the nucleotides are present.}
}
\value{
TRUE or FALSE
}
\description{
Check if a character vector contains either at least one, or all, of the four DNA nucleotides (A C G T).
}
