% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.overlap.R
\name{clusterIntervals}
\alias{clusterIntervals}
\title{Clusters intervals

Merges overlapping intervals and returns a data frame with names of merged intervals and start and end of merged intervals.}
\usage{
clusterIntervals(data)
}
\arguments{
\item{data}{two column data frame with limits of each interval held on a separate row}
}
\value{
Three column data frame with the set of nonoverlapping intervals formed after merging overlapping intervals of data; the first column holds the comma-separated rownames of data that were merged; columns two and three hold the start and end positions of the post-merge interval
}
\description{
Clusters intervals

Merges overlapping intervals and returns a data frame with names of merged intervals and start and end of merged intervals.
}
