% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.DNAStringSet.R
\name{collapse.DNAStringSet}
\alias{collapse.DNAStringSet}
\title{Collapse DNAStringSets

Collapse a list of DNAStringSet objects into a single DNAStringSet object.}
\usage{
\method{collapse}{DNAStringSet}(input.seqs, use.seqnames = T, use.set.names = T)
}
\arguments{
\item{input.seqs}{A list of DNAStringSet objects.}

\item{use.seqnames}{Should the names of the input DNAStringSets be incorporated into the seqnames of the collapsed DNAStringSet. Default T.}

\item{use.set.names}{Should the names of the input DNAStringSets be incorporated into the seqnames of the collapsed DNAStringSet. Default T.}
}
\value{
A single DNAStringSet object containing the sequences from all of the DNAStringSet objects in input.seqs. Sequence names are determined by use.seqnames and use.set.names parameters.
}
\description{
Collapse DNAStringSets

Collapse a list of DNAStringSet objects into a single DNAStringSet object.
}
