% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.low.support.nodes.R
\name{collapse.low.support.nodes}
\alias{collapse.low.support.nodes}
\title{Collapse Low Supported Nodes}
\usage{
collapse.low.support.nodes(rooted.tree, support.threshold = 80)
}
\arguments{
\item{rooted.tree}{Rooted input phylogeny.}

\item{support.threshold}{Nodes with support less than this value are collapsed into a polytomy. Default is 80, but care should be taken for this value.}
}
\value{
Returns a phylogeny with the low support nodes collapsed into polytomies.
}
\description{
Given an input phylogeny, returns a phylogeny with low support nodes (below a threshold value of support) collapsed into polytomies.
}
