% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column.split.R
\name{column.split}
\alias{column.split}
\title{Split Columns by Delimiter}
\usage{
column.split(x, split = " ", ncol = "auto", byrow = T)
}
\arguments{
\item{x}{Input matrix and column to be split}

\item{split}{Character to delimit columns by}

\item{ncol}{number of columns in final matrix (default is to determine automatically)}

\item{byrow}{Logical indicating if values in the result matrix should be filled by rows and then columns}
}
\value{
Output matrix
}
\description{
Split a matrix column into multiple columns by a delimiter
}
