% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexity.R
\name{complexityPlot}
\alias{complexityPlot}
\title{Plot sliding window DNA complexity

Plots sequence complexity of an input DNA sequence along sliding windows and a random sequence of the same length.
Black line shows complexity of the input DNA sequence for a sliding window with windowsize=10 and windowshift=5
Red line shows complexity of the input DNA sequence for a sliding window with windowsize=200 and windowshift=100
Blue line shows complexity of the randomly generated DNA sequence using a sliding window with windowsize=200 and windowshift=100
Note: I may change this to a ggplot and provide options to control sliding windows}
\usage{
complexityPlot(dna, wsize = 1)
}
\arguments{
\item{dna}{Character string, or a DNAString object, or a length 1 DNAStringSet object as defined by the Biostrings package.}

\item{wsize}{Number that controls windowsizes; default window sizes of the larger windows (indicated by red and blue lines) are multiplied by this number.}
}
\description{
Plot sliding window DNA complexity

Plots sequence complexity of an input DNA sequence along sliding windows and a random sequence of the same length.
Black line shows complexity of the input DNA sequence for a sliding window with windowsize=10 and windowshift=5
Red line shows complexity of the input DNA sequence for a sliding window with windowsize=200 and windowshift=100
Blue line shows complexity of the randomly generated DNA sequence using a sliding window with windowsize=200 and windowshift=100
Note: I may change this to a ggplot and provide options to control sliding windows
}
