% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatenate.alignments.R
\name{concatenate.alignments}
\alias{concatenate.alignments}
\title{Concatenate All Alignments in Directory}
\usage{
concatenate.alignments(
  alignment.folder,
  type = "DNA",
  seqs.format = "phylip",
  alignment.output = NA,
  partition.output = NA,
  show.progress = T,
  istart = NA,
  iend = NA,
  only.variable.sites = F
)
}
\arguments{
\item{alignment.folder}{Path to the directory containing the input set of alignments that should be concatenated.}

\item{type}{Type of sequences (default = "DNA"). Alternatively "AA" for Amino Acid sequences.}

\item{seqs.format}{Format of input sequences (default = "phylip"). No other formats currently implemented.}

\item{alignment.output}{Where to save the concatenated alignment (default value NA does not save the concatenated alignment).}

\item{partition.output}{Where to save the partition file for the concatenated alignment (default value NA does not create a partition file).}

\item{show.progress}{Print progress to screen (default TRUE)}

\item{istart}{Which alignment to start at (default is NA, meaning all alignments in alignment.folder should be concatenated)}

\item{iend}{Which alignment to end at (default is NA, meaning all alignments in alignment.folder should be concatenated)}

\item{only.variable.sites}{Should invarient sites be removed? Default is FALSE.}
}
\value{
DNAStringSet or AAStringSet object containing the concatenated sequences in the order of list.files(alignment.folder).
}
\description{
Concatenate all phylip-formatted DNA or AA alignments that are stored in separate input files in the same input directory, and optionally create a partition file.
}
