% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatenateDNAStringSets.R
\name{concatenateDNAStringSets}
\alias{concatenateDNAStringSets}
\title{Concatenate DNAStringSets}
\usage{
concatenateDNAStringSets(...)
}
\arguments{
\item{...}{Two or more DNAStringSet objects separated by commas. All taxa must be present in all input DNAStringSets. All sequences in each DNAStringSet must have the same length and are assumed to represent homologous sites.}
}
\value{
A DNAStringSet containing the concatenated sequences in the order in which input DNAStringSets were supplied.
}
\description{
Concatenate multiple DNA sequence alignments stored in DNAStringSet objects.
}
