% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/congruency.network.R
\name{congruency.network}
\alias{congruency.network}
\title{Phylogenetic Congruency Network from Congruency Matrix}
\usage{
congruency.network(congruency.matrix, plot.network = T)
}
\arguments{
\item{congruency.matrix}{The matrix object generated by running multiple.congruency.test, which is used as input in this function.}

\item{plot.network}{Whether or not to plot the network (default is true).}
}
\value{
An igraph network object. Each node is a tree and edges connect trees that have congruent topologies.
}
\description{
Plot the result of the function multiple.congruency.test as a network to summarize topological discordance among phylogenies.
}
