% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countChars.R
\name{countChars}
\alias{countChars}
\title{Count character occurences}
\usage{
countChars(x, pattern)
}
\arguments{
\item{x}{character string to search within}

\item{pattern}{the character or character string to be counted in x}
}
\value{
integer of the number of occurrence of pattern in x
}
\description{
Count the number of occurrences of a character in a character string
}
