% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{datasets}
\alias{datasets}
\title{REEs Package Datasets}
\usage{
datasets(x = NULL)
}
\arguments{
\item{x}{A character string, character vector, or NULL (default). If NULL, a character vector of possible values to use for this argument is returned.}
}
\value{
Character vector containing the URLs of the fasta-formatted genomes for the species used to design REEs in the SnakeCap study. The names attribute holds the corresponding species names.
}
\description{
This function is used to return data associated with the REEs package
}
