% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.alignment.R
\name{filter.alignment}
\alias{filter.alignment}
\title{Filter Alignment}
\usage{
\method{filter}{alignment}(
  alignment,
  mdt = 1,
  treat.ambiguous.as.missing = F,
  min.allele.freqs.dna = c(0, 0, 0, 0),
  min.allele.freqs.aa = rep(0, 21),
  keep.specific = NA,
  drop.nodata.samples = F
)
}
\arguments{
\item{alignment}{Input alignment of class DNAStringSet, DNAMultipleAlignment, or DNAbin}

\item{mdt}{Missing data threshold: maximum fraction of missing data allowed at a site (default = 1)}

\item{treat.ambiguous.as.missing}{Logical indicating if ambiguous characters should be treated as missing data (default is FALSE).}

\item{min.allele.freqs.dna}{Numerical vector of length n, indicating the minimum frequencies of the most frequent, second-most frequent, third-most frequent, and fourth-most frequent allele required to return a value of TRUE.
Additionally, the length of the input vector sets the maximum number of allele types allowed. The number of non-zero values sets the minimum number of allele types required.
Default is c(0,0,0,0), meaning that up to four different DNA nucleotides are allowed, and that none of these need to be present for the function to return a value of TRUE for the site.}

\item{min.allele.freqs.aa}{Numerical vector. Values determine the minimum frequency of the most frequent, second-most frequent, ..., and n-most frequent allele type (amino acid) required to return a value of TRUE.
Additionally: (1) the length of min.freqs determines the maximum number of different AAs allowed or else function returns FALSE.
(2) the number of non-zero elements in min.freqs determines the minimum number of different AAs required or else the function returns FALSE.
Default value is a vector of zeros with length 21, meaning that up to 21 allele types (the 20 amino acids and stop codon) are allowed, and that none of these AAs need to be present for the function to return a value of TRUE.}

\item{keep.specific}{Providing a numeric range or set allows you to extract specific sites (default NA).}

\item{drop.nodata.samples}{Drops an individual if it doesnt have any data for the sites that meet the other criteria (default FALSE).}
}
\value{
A filtered DNA or AA alignment
}
\description{
Filters individuals or sites from an DNA or AA alignment based on user-defined thresholds of percent missing or ambiguous data per site, variation per site, frequency of common alleles per site, frequency of rare alleles per site, or specific sites (e.g., every second, third, etc.)
Running this function with the default values for arguments returns the input alignment unchanged.
}
