% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.gff.R
\name{filter.gff}
\alias{filter.gff}
\title{Filter Feature Annotation Table}
\usage{
filter.gff(input.gff, output.gff = NULL, feature.type, min.length = 0)
}
\arguments{
\item{input.gff}{Either a path to input feature annotation table (gff format) or an object of class data.table. See "ref_Thamnophis_sirtalis-6.0_top_level_JLW.gff3" as an example input.}

\item{output.gff}{Where to write the output (filtered) annotation table. Default is NULL, in which case the output is not written. See "CDS_ref_Thamnophis_sirtalis-6.0_top_level_JLW_longer120bp.gff3" as an example output.}

\item{feature.type}{Type of sequences to keep. Options include all unique values in the "feature" column of input.gff. The usually include the following: "region", "gene", "mRNA", "exon", "CDS", "lnc_RNA", "cDNA_match", "transcript", "pseudogene", "tRNA", "V_gene_segment", "C_gene_segment".}

\item{min.length}{Minimum nucleotide length to include a feature in the output annotation table. Default is 120.}

\item{write.table}{If TRUE, then the filtered feature table is written to the value of}
}
\value{
Writes the filtered annotation table to the path indicated by the output.gff argument. Additionally returns a data.table object.
}
\description{
Filters a GFF-format feature annotation table to include only features of a particular feature type and minimum feature length.
The output of this function is a GFF file containing the filtered an annotation table, which can be used as input to the functions get.loci.from.annotationTable or get.exome.from.annotationTable
}
