% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.md.R
\name{filter.md}
\alias{filter.md}
\title{Test Missing Data

Test if a character vector contains. This function is used internally in the the filter.alignment function.}
\usage{
filter.md(alignment.site, missing.data.threshold, type, ambiguous = F)
}
\arguments{
\item{alignment.site}{Character vector corresponding to a site in a DNA or AA alignment}

\item{missing.data.threshold}{Value between 0 and 1 and determines the maximum fraction of missing characters in the input vector for the function to return TRUE.}

\item{type}{Either "DNA" or "AA"}

\item{ambiguous}{Logical indicating if ambiguous characters should be treated as missing data (default is FALSE)}
}
\value{
Returns TRUE if no missing data, or if the fraction of missing data is less than the value of missing.data.threshold argument; otherwise returns FALSE.
}
\description{
Test Missing Data

Test if a character vector contains. This function is used internally in the the filter.alignment function.
}
