% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.var.aa.R
\name{filter.var.aa}
\alias{filter.var.aa}
\title{Filter by AA Variation}
\usage{
filter.var.aa(alignment.site, min.freqs = rep(0, 21))
}
\arguments{
\item{alignment.site}{Input character vector representing AA data at one site of an alignment.}

\item{min.freqs}{Numerical vector. Values determine the minimum frequency of the most frequent, second-most frequent, ..., and n-most frequent allele type (amino acid) required to return a value of TRUE.
Additionally: (1) the length of min.freqs determines the maximum number of different AAs allowed or else function returns FALSE.
(2) the number of non-zero elements in min.freqs determines the minimum number of different AAs required or else the function returns FALSE.
Default value is a vector of zeros with length 21, meaning that up to 21 allele types (the 20 amino acids and stop codon) are allowed, and that none of these AAs need to be present for the function to return a value of TRUE.}
}
\value{
Logical indicating if the site meets all of the input requirements for amino acid variation.
}
\description{
This function is used internally in the function filter.alignment to test if an alignment site meets threshold levels of amino acid variation.
Default values return TRUE for any input character vector.
}
