% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.var.dna.R
\name{filter.var.dna}
\alias{filter.var.dna}
\title{Filter by DNA Variation}
\usage{
filter.var.dna(alignment.site, min.freqs = c(0, 0, 0, 0))
}
\arguments{
\item{alignment.site}{Input character vector representing DNA data at one site of an alignment.}

\item{min.freqs}{Numerical vector of length n, indicating the minimum frequencies of the most frequent, second-most frequent, third-most frequent, and fourth-most frequent allele required to return a value of TRUE.
Additionally, the length of min.freqs sets the maximum number of allele types required, and the number of non-zero elements in min.freqs sets the minimum number of allele types required.
Default is c(0,0,0,0), meaning that up to four allele types are allowed (the four DNA nucleotides), and that none of these need to be present for the function to return a value of TRUE.}
}
\value{
Logical indicating if the site meets all of the input requirements for nucleotide variation.
}
\description{
This function is used internally in the function filter.alignment to test if a DNA alignment site is meets threshold levels of variation.
Default values return TRUE for any input character vector.
}
