% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.orf.R
\name{find.orf}
\alias{find.orf}
\title{Find Open Reading Frames}
\usage{
find.orf(input.seq, codons = F, min.size = 80)
}
\arguments{
\item{input.seq}{DNAStringSet object containing one sequence.}

\item{codons}{Whether or not the function should return a table with the start and stop positions and frame (relative to sequence) of all potential stop codons. Default is FALSE.}

\item{min.size}{Minimum open reading frame size to consider (default = 80).}
}
\value{
A data frame holding one of the following:
(1) Three column data frame indicating position and frame of each possible stop codons.
(2) Four column data frame indicating position, nucleotide length, and frame relative to input sequence, of all open reading frames at least as long as specified by the min.size argument.
}
\description{
Given an input DNA sequence, returns either a table of stop codon positions and frame relative to input sequence, or a table of open indicating the positions, lengths, and frames of open reading frames.
}
