% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.all.quartets.R
\name{get.all.quartets}
\alias{get.all.quartets}
\title{Get All Quartets}
\usage{
get.all.quartets(ref.tree, alt.tips = NULL, support.scaler = 1)
}
\arguments{
\item{ref.tree}{Newick format phylogenetic tree}

\item{alt.tips}{Alternative set of tips (characterized by tip labels) to create quartets for. Default is NULL, meaning that all tips in the input tree are all allowed in quartets and all possible quartets are included in output. If alt.tips is set to a value other than NULL, any quartets containing tips not in the alt.tips are filtered from the output. It is often useful to set alt.tips = unique.tiplabels(multiPhylo), where multiPhylo is an object of class multiPhylo that holds a set of gene trees.}

\item{support.scaler}{Value to multiply the support values by (default is 1). It might be useful to scale the support values of some trees if comparing trees with different methods of support (e.g., if comparing Bayesian and ML trees). I haven't experimented with this yet.}
}
\value{
multiPhylo object containing the set of all quartets for an input tree and a set of tip labels.
}
\description{
Get the set of all 4-taxon sub-trees (quartets) of an input tree
}
