% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitogenomes.R
\name{get.mitogenome}
\alias{get.mitogenome}
\title{Get mitochondrial genome from reads}
\usage{
get.mitogenome(
  sampleName,
  read1,
  read2 = NULL,
  reads.merged = NULL,
  referencePATH,
  geneFilePATH,
  out.dir,
  bbmapPATH = "bbmap.sh'",
  cap3PATH = "cap3",
  spadesPATH = "spades.py",
  pblatPATH = "pblat"
)
}
\arguments{
\item{sampleName}{Name of the sample}

\item{read1}{Path to fasta file containing cleaned, unmerged (unmeargeable) read1 reads. This file has '_READ1_unmerged' in the filename if it was generated with the SnakeCap pipeline.}

\item{read2}{Path to fasta file containing cleaned, unmerged (unmeargeable) read2 reads. This file has '_READ2_unmerged' in the filename if it was generated with the SnakeCap pipeline.}

\item{reads.merged}{Path to fasta file containing the cleaned, merged reads. This file has 'singletons' in the filename if it was generated with the SnakeCap pipeline.}

\item{referencePATH}{Path to fasta file containing mitogenomes of many snakes.}

\item{geneFilePATH}{Path to fasta file containing genes for one of the mitogenomes.}

\item{out.dir}{Path to directory where output should be saved.}

\item{bbmapPATH}{Path to bbmap bash script. Default "bbmap.sh", in which case bbmap. must be on your system path.}

\item{cap3PATH}{Path to cap3 executable. Default "cap3", in which case cap3 must be on your system path.}

\item{spadesPATH}{Path to spades python executable. Default "spades.py", in which case spades must be on your system path.}

\item{pblatPATH}{Path to pblat executable. Default "pblat", in which case pblat must be on your system path.}
}
\value{
DNAStringSet; mitogenome is written to out.dir
}
\description{
Get mitochondrial genome from reads
}
