% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.seqs.from.blastTable.R
\name{get.seqs.from.blastTable}
\alias{get.seqs.from.blastTable}
\title{Get Sequences From BLAST Table

Extracts the sequences defined in an input table (a blast table with NCBI format 6) from an input genome. The input genome should be the same genome used as the subject in the blast table.
The quality/completeness of the output sequence dataset depends on the quality/completeness of the blast search.
The functions get.exome.from.blastTable and get.UCEs.from.blastTable are the same as this one, but applied to a specific type of sequences in the blast table. This function could supercede those functions.}
\usage{
get.seqs.from.blastTable(input.blastTable, input.seqs, output.path = NULL)
}
\arguments{
\item{input.blastTable}{Path to the blast table containing best matches to query sequences found in the input genome (subject in blast search). This input table is the output of the reportBestMatches function.}

\item{input.seqs}{Path to the input genome.}

\item{output.path}{Directory where to save the extracted sequences.}
}
\value{
Extracted sequences written to a file.
}
\description{
Get Sequences From BLAST Table

Extracts the sequences defined in an input table (a blast table with NCBI format 6) from an input genome. The input genome should be the same genome used as the subject in the blast table.
The quality/completeness of the output sequence dataset depends on the quality/completeness of the blast search.
The functions get.exome.from.blastTable and get.UCEs.from.blastTable are the same as this one, but applied to a specific type of sequences in the blast table. This function could supercede those functions.
}
