% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.seqs.from.gff.R
\name{get.seqs.from.gff}
\alias{get.seqs.from.gff}
\title{Get Sequences from GFF Table}
\usage{
get.seqs.from.gff(
  input.seqs,
  input.gff,
  output.path = NULL,
  additional.ID = NULL
)
}
\arguments{
\item{input.seqs}{Character string with either the local path or URL to fasta formatted sequences. File can be .gz compressed.}

\item{input.gff}{One of the following: An object of class data table, data.frame, or character matrix, with columns identical to those used in Generic Feature Format (GFF) tables; or, a character string with local file path to either a GFF file or table with columns identical to those used in GFF files; or, a character string with URL path to a GFF file.}

\item{output.path}{Character string with path where sequences should be saved (default is NULL).}

\item{additional.ID}{One of the following: NULL (the default, and usually what you want to use); or, a character string with local filepath to a two-column table with contig names in GFF and sequence file, respectively. Usually, contig names are already the same in GFF and sequence files, but this argument can be useful if you renamed contigs in either the GFF or sequence file.}
}
\value{
DNAStringSet object with sequences corresponding to unique ranges in input.gff; sequences are written to output.file in fasta format if output.path is a character string.
}
\description{
Identical to the function get.exome.from.gff, but used more generally.
Note to self: I used this function to extract the MHC loci from the Thamnophis sirtalis genome.
}
