% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ncbi_sequences.R
\name{get_ncbi_sequences}
\alias{get_ncbi_sequences}
\title{Download NCBI Sequences}
\usage{
get_ncbi_sequences(
  outfile = NULL,
  input.seqs = NULL,
  accessionList,
  startList = 1,
  endList,
  strandList = "1",
  db = "nuccore",
  rettype = "fasta",
  retmode = "text",
  if.outside.range = "partial",
  trim.ambiguous = FALSE
)
}
\arguments{
\item{outfile}{Filepath where sequences should be saved.}

\item{input.seqs}{Either NULL (default but not preferred usage), or a character string indicating either a local filepath or URL path to the input sequences in fasta format. The URL character string is often the preferred method.}

\item{accessionList}{Character vector of accession numbers.}

\item{startList}{Numeric vector of starting positions (default = 1)}

\item{endList}{Numeric vector of end positions.}

\item{strandList}{Single character or character vector indicating the strand of the sequences (default = "1"). If a single character supplied, then the same strand will be downloaded for all sequences. If length(strandList)>1, then length(strandList) should equal length(accessionList). Set strandList = "1" to download forward direction sequences; strandList = "2" to download reverse complement of sequences.}

\item{db}{NCBI database to download sequences from (default is "nuccore"). I have not tested this function for other databases.}

\item{rettype}{Format to save sequences (default is "fasta"). Alternatively, "gb" (= download sequences in GenBank flatfile format), "native" (= XML format), "acc" (only download accession numbers), "seqid" (download SeqID string), ft" (download feature tables).}

\item{retmode}{String indicating the mode to download sequences (default = "text"). Don't change from default.}

\item{if.outside.range}{Ignored if input.seqs is NULL. Otherwise, one of the following character strings: "partial" (the default), "drop", or "error". If partial, then the overlapping portion of sequences are returned if the requested ranges  only partially overlap with the contig. If "error", an error is returned indicating which requested sequences are outside the range of the contig. If "drop", then only requested sequences that are entirely contained within a contig are returned.}

\item{trim.ambiguous}{Whether or not to trim N strings from the ends of sequences after downloading (default = FALSE). This argument is ignored if input.seqs is NULL.}
}
\value{
A file containing the downloaded sequences.
}
\description{
Download specific regions of a set of sequences from NCBI.
}
