% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include.accession.region.gb.R
\name{include.accession.region.gb}
\alias{include.accession.region.gb}
\title{Add Region Info to GenBank Flatfile}
\usage{
include.accession.region.gb(gb.filename)
}
\arguments{
\item{gb.filename}{}
}
\value{
A temporary file that is a modified version of the input file that is read into R when running the read.gb function.
}
\description{
This function is implemented within the read.gb function, and may not have other stand-alone applications.
The function creates a temporary file (.tmp) containing a modified version of the input GenBank flatfile, in which the region information (i.e., start and end info)
is included as part of the accession name. This modification allows the region info to be acccessed when using the AA.from.gb function.
}
