% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mafft.install.R
\name{initialize_REEs}
\alias{initialize_REEs}
\title{Initialize REEs}
\usage{
initialize_REEs()
}
\value{
Currently NULL. Soon this function will return a data frame showing where REEs is looking for the programs it depends on.
}
\description{
This function installs some dependency programs from source to a location where REEs knows where to find it.
}
