% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect.all.R
\name{intersect.all}
\alias{intersect.all}
\title{Intersect of all sets}
\usage{
intersect.all(x)
}
\arguments{
\item{x}{Input list of vectors}
}
\value{
Output vector of elements shared among x
}
\description{
Get the intersect set of two or more sets
}
