% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.overlap.R
\name{is.overlap}
\alias{is.overlap}
\title{Test if Two Ranges Overlap

This function takes as input two integer vectors. The min and max values of each vector represent the lower and upper limits of each range, respectively. Returns TRUE if input ranges overlap, otherwise FALSE.}
\usage{
is.overlap(r1, r2)
}
\arguments{
\item{r1}{Integer vector with min and max values representing limits of first range}

\item{r2}{Integer vector with min and max values representing limits of second range}
}
\value{
Logical indicating whether or not r1 and r2 overlap
}
\description{
Test if Two Ranges Overlap

This function takes as input two integer vectors. The min and max values of each vector represent the lower and upper limits of each range, respectively. Returns TRUE if input ranges overlap, otherwise FALSE.
}
