% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.overlap.R
\name{is.subset}
\alias{is.subset}
\title{Test if subset

Test if a range of numbers is a subset of another range of numbers.}
\usage{
is.subset(r1, r2)
}
\arguments{
\item{r1}{Integer vector with min and max values representing limits of first range}

\item{r2}{Integer vector with min and max values representing limits of second range}
}
\value{
Logical indicating whether or not r1 is a non-equivalent subset of r2.
}
\description{
Test if subset

Test if a range of numbers is a subset of another range of numbers.
}
