% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.wholenumber.R
\name{is.wholenumber}
\alias{is.wholenumber}
\title{Whole number test}
\usage{
is.wholenumber(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{input number}

\item{tol}{tolerance?}
}
\value{
TRUE if x is a whole number, otherwise FALSE
}
\description{
Test if an input number is a whole number
}
