% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.gff.R
\name{load.gff}
\alias{load.gff}
\title{Read a General Feature Format (GFF) file}
\usage{
load.gff(input, local = TRUE)
}
\arguments{
\item{input}{A character string giving the path to the GFF file. If input is a URL then local must be FALSE.}

\item{local}{Whether or not input is a path to a local file (default is TRUE). If set to FALSE, then the function treats input as URL.}
}
\value{
A data table object containing the GFF feature rows. Incomplete feature rows are dropped.
}
\description{
This function reads a General Feature Format (GFF) file. If the GFF is a local file, then this function is simply a wrapper for the function read.gff() in the package ape.
If the GFF is stored in an open source online data repository (e.g. NCBI) this function first downloads a temporary local copy of the file.
See http://gmod.org/wiki/GFF3 for info about GFF files.
}
