% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PostsequencingPipeline.R
\name{lociAlignment04}
\alias{lociAlignment04}
\title{Step 4: Align Captured Loci}
\usage{
lociAlignment04(workdir = getwd(), targets.path, istart = 1, iend = NULL)
}
\arguments{
\item{workdir}{Path to the directory containing the 'Processed_Samples' folder used in steps 1–3 of SnakeCap. Default is the current directory.}

\item{targets.path}{Path to the file containing the loci that were targetted for capture, and from which probes were designed.}

\item{istart}{Number indicating the first locus to process, among the set of unique contigs that do not match other targets. Default 1.}

\item{iend}{Number indicating the last locus to process, among the set of unique contigs that do not match other targets. Default is NULL, in which case all loci including and after istart will be processed.}
}
\value{
NULL # Later this function will return a data frame with the paths to output files
}
\description{
This function performs step 4 of the post-sequencing pipeline used for SnakeCap
MAFFT is used for multiple sequence alignment with the settings '--localpair --maxiterate 1000 --adjustdirection --quiet --op 3 --ep 0.123 --thread 6', and using a jacknifing approach to find the the set of putatively orthologous contigs among individuals.
The input sequences should be the fasta files generated in step 3 that are named '.*contigs-SingleTargetMatches.fa'
}
