% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maf2fasta.R
\name{maf2fasta}
\alias{maf2fasta}
\title{Convert Multiple Alignment Format to Fasta

For each MAF alignment block, this function writes a fasta alignment file}
\usage{
maf2fasta(
  maf,
  outdir,
  samples = NULL,
  minwidth = 100,
  minsamples = 4,
  silent = F
)
}
\arguments{
\item{maf}{Path to the input MAF file}

\item{outdir}{Path to directory where output fasta files should be written}

\item{samples}{Names of samples to keep. Default NULL means all samples are kept.}

\item{minwidth}{Number indicating the minimum alignment width to retain an alignment. Default 100.}

\item{minsamples}{Number indicating the minimum number of individuals in an alignment to keep the alignment. Default 4.}

\item{silent}{Logical indicating whether or not to omit printing to screen. Default TRUE.}
}
\value{
NULL; alignments are written to outdir
}
\description{
Convert Multiple Alignment Format to Fasta

For each MAF alignment block, this function writes a fasta alignment file
}
