% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mafft.install.R
\name{mafft.install}
\alias{mafft.install}
\title{Install MAFFT}
\usage{
mafft.install(install.loc = "auto", source = F)
}
\arguments{
\item{install.loc}{Path where MAFFT should be installed, or "auto" (Default), which installs MAFFT to blast-mafft subdirectory of REEs directory, or "PATH", which installs to "/usr/local" if it is writeable.}
}
\value{
NULL. MAFFT is installed to install.loc
}
\description{
This function installs MAFFT from source to the REEs directory or a user-specified location.
This code only works if you Command Line Developer Tools installed.
}
