% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mafft.R
\name{mafft}
\alias{mafft}
\title{MAFFT Multiple sequence alignment}
\usage{
mafft(x, param = "--auto", mafft.path = NULL, return.as = "XStringSet")
}
\arguments{
\item{x}{Object of class XStringSet (e.g., DNAStringSet) that contains the sequences to be aligned.}

\item{param}{Character string containing the mafft command line parameters.}

\item{mafft.path}{Either NULL (default) or a character string specifying the path to the mafft executable. If NULL, the mafft executable must be on your system PATH. This is very likely to be the case because by default mafft installs to usr/local/bin.
A possible exception is when mafft is used on a community cluster. In this case, specify the path to the mafft executable instead of using "auto".}

\item{return.as}{Of the following: "XStringSet" (default) or "MultipleAlignment"}
}
\value{
Either an object of class XStringSet (if return.as= "XStringSet") or DNAMultipleAlignment (or RNA MultipleAlignment or AAMultipleAlignment) if return.as = MultipleAlignment (see BioStrings).
}
\description{
Wrapper for MAFTT. Perform. multiple sequence alignment on DNA, RNA, or AA sequences.
This function was lightly modified from the mafft function written by Michael Hahsler and Anurag Nagar.
}
