% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.consensus.R
\name{make.consensus}
\alias{make.consensus}
\title{Get Consensus Sequence}
\usage{
make.consensus(
  input.alignment,
  method = c("majority", "threshold", "IUPAC", "profile"),
  threshold = 0.6,
  warn.non.IUPAC = FALSE,
  type = "DNA"
)
}
\arguments{
\item{input.alignment}{Input DNA alignment of class DNAStringSet or DNAMultipleAlignment}

\item{method}{One of the following methods to determine consensus sequences: "majority","threshold","IUPAC","profile". These are defined in seqinr package.}

\item{threshold}{When method = threshold, this is the minimum relative frequency for a character to be returned as the consensus character (default = 0.6).}

\item{warn.non.IUPAC}{Warn if sequence data includes non-IUPAC standard characters. Default is FALSE.}

\item{type}{Only available option is "DNA"}
}
\value{
DNAStringSet object containing the consensus sequence of the input DNA alignment
}
\description{
This function returns the consensus sequence of an input DNA alignment; it is a slightly modified version of the consensus() function in the seqinr package. The only difference is that this function can additionally take as input alignments of class DNAStringSet and DNAMultipleAlignments, and the result is an alignment of class DNAStringSet.
}
