% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.partitioned.alignment.R
\name{make.partitioned.alignment}
\alias{make.partitioned.alignment}
\title{Make Partitioned Alignments

Given an input DNA alignment and a reference CDS sequence for the locus in the input alignment, this function generates up to nine output alignments, including three partitioned DNA alignments, five DNA alignments containing a subset of the input DNA sites, and one amino acid alignment for the CDS region(s) of the input alignment.}
\usage{
make.partitioned.alignment(
  input.path,
  output.dir,
  TargetCDS.path,
  steps = "1111111",
  mafft.params = " --auto --adjustdirection --nwildcard --op 3 --ep 0.123 --quiet ",
  old.names = NA,
  new.names = NA,
  drop.reference = F,
  ith.locus.start = 1,
  ith.locus.end = "all",
  locus.names.omit = NULL,
  AA.pdist.drop.thresh = 0.5,
  trimto = NULL
)
}
\arguments{
\item{input.path}{Folder containing fasta-formatted input DNA alignments.}

\item{output.dir}{Where to save alignments (alignments for each type of data will be in different subdirectory)}

\item{TargetCDS.path}{Full path to the fasta file containing only the CDS sequences of target loci (from which probes were designed). Sequence names must have the following format: "GeneName_TargetCDS_of_TargetLocusName_AnyAdditionalIformation", where GeneName and TargetLocusName are replaced with the actual names, and AnyAdditionalIformation can be a string of any characters}

\item{steps}{Character string with nine character, each 0 or 1, indicating which steps of the function to do (1) or skip (0). Used for debugging.}

\item{mafft.params}{Character string with parameters passed to MAFFT. Default is " --auto --adjustdirection --nwildcard --op 3 --ep 0.123 --quiet "}

\item{old.names}{Character vector of old sample names that should be changed to those defined in new.names parameter (default is NA, meaning don't change names).}

\item{new.names}{Character vector of new sample names to use instead of corresponding name in old.names vector (default is NA, meaning don't change names).}

\item{drop.reference}{Whether or not to include the reference sequence (from which baits were designed for a particular locus) in the alignment written to file (default is FALSE).}

\item{ith.locus.start}{First locus in input.path to process (default is 1).}

\item{ith.locus.end}{Last locus in input.path to process (default is "all", when means process until no more loci to process).}

\item{locus.names.omit}{Names of loci to skip (default is NULL).}

\item{AA.pdist.drop.thresh}{Maximum mean pairwise p-distance for AA sequence of an individual (translated from CDS regions of input DNA sequences) to keep the individual in AA or CDS-containing DNA alignments.}

\item{trimto}{Optional character string with names of sequences in input alignments whose combined range should be used to trim the alignments. Default is NULL.}

\item{ref.type}{Indicates that the input alignments are DNA sequences (default is "DNA").}
}
\value{
NULL; writes up to nine different fasta-formatted sequence alignments for each input DNA alignment in input.path, and partition files partitioned alignments. The nine output alignment types are:
(1) Partitioned version of the input DNA alignment. The alignment is partitioned into (if present): upstream noncoding DNA, each codon position (within CDS regions), and downstream noncoding DNA.
(2) Alignment containing only the CDS regions, which are partitioned by codon position.
(3) Alignment containing only the upstream noncoding DNA.
(4) Alignment containing only the downstream noncoding DNA.
(5) Alignment containing only noncoding DNA, partitioned into upstream and downstream regions.
(6) Alignment containing only the first codon positions of CDS regions.
(7) Alignment containing only the second codon positions of CDS regions.
(8) Alignment containing only the third codon positions of CDS regions.
(9) Alignment containing the amino acid sequence for the translated CDS region.
}
\description{
Make Partitioned Alignments

Given an input DNA alignment and a reference CDS sequence for the locus in the input alignment, this function generates up to nine output alignments, including three partitioned DNA alignments, five DNA alignments containing a subset of the input DNA sites, and one amino acid alignment for the CDS region(s) of the input alignment.
}
