% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{makeBlastDB}
\alias{makeBlastDB}
\title{Make Local Blast Database

Wrapper for the NCBI makeblastdb command. This function makes an NCBI blast database (which is needed to run blast against).}
\usage{
makeBlastDB(makeblastdb.path, subject.path)
}
\arguments{
\item{makeblastdb.path}{Path to makeblastdb (included in the bin folder of blast)}

\item{subject.path}{Path to the file containing the sequences to include in the database}
}
\value{
Creates a blast database file.
}
\description{
Make Local Blast Database

Wrapper for the NCBI makeblastdb command. This function makes an NCBI blast database (which is needed to run blast against).
}
\examples{

### Load GFF table from NCBI repository.
Thamnophis.sirtalis_GFF.url <- "https://ftp.ncbi.nlm.nih.gov/genomes/all/GCF/001/077/635/GCF_001077635.1_Thamnophis_sirtalis-6.0/GCF_001077635.1_Thamnophis_sirtalis-6.0_genomic.gff.gz"
Thamnophis.sirtalis_GFF     <- load.gff(input=Thamnophis.sirtalis_GFF.url,local=F)

# Filter Thamnophis.sirtalis_GFF to only include CDS features with length at least 120bp
Thamnophis.sirtalis_GFF_CDS_longer120bp <- filter.gff(input.gff=Thamnophis.sirtalis_GFF,feature.type="CDS",min.length=120)

### Use get.seqs.from.gff to extract the sequences for the loci in the filtered GFF
Thamnophis.sirtalis_genome.path <- "https://ftp.ncbi.nlm.nih.gov/genomes/all/GCF/001/077/635/GCF_001077635.1_Thamnophis_sirtalis-6.0/GCF_001077635.1_Thamnophis_sirtalis-6.0_genomic.fna.gz"
Thamnophis.sirtalis_exome <- get.seqs.from.gff(input.seqs=Thamnophis.sirtalis_genome.path,input.gff=Thamnophis.sirtalis_GFF_CDS_longer120bp)

### Use tblastx to return up to 50 matches for each of exon (only first two exons in this example) of Thamnophis exome in Crotalus horridus genome
test.query    <- Thamnophis.sirtalis_exome[1:2]
test.subject  <- "https://ftp.ncbi.nlm.nih.gov/genomes/all/GCA/001/625/485/GCA_001625485.1_ASM162548v1/GCA_001625485.1_ASM162548v1_genomic.fna.gz"
test.50hits   <- blast(method="tblastx",subject=test.subject,query=test.query)
}
