% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat.strsplit.R
\name{mat.strsplit}
\alias{mat.strsplit}
\title{Split character string to matrix}
\usage{
mat.strsplit(x, split = " ", byrow = T)
}
\arguments{
\item{x}{Input character string to be split}

\item{split}{Character delimeter to split by (default is to split at spaces)}

\item{byrow}{Whether values in the output matrix should be filled by rows (default TRUE) or columns}
}
\value{
Output matrix
}
\description{
Split a string into a matrix containing the splitted strings
}
