% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.overlap.R
\name{mergedRange}
\alias{mergedRange}
\title{Return range of merged intervals if they overlap

This function takes as input two integer vectors. The min and max values of each vector represent the lower and upper limits of each range, respectively. Returns the range of the merged interval if r1 and r2 overrlap, otherwise NA.}
\usage{
mergedRange(r1, r2, ifFALSE = NA)
}
\arguments{
\item{r1}{Integer vector with min and max values representing limits of first range}

\item{r2}{Integer vector with min and max values representing limits of second range}

\item{ifFALSE}{What to return when r1 and r2 are nonoverlapping. Default NA; ifFALSE=1 returns r1; ifFALSE=2 returns r2; ifFALSE=12 returns a length 2 list with r1 and r2}
}
\value{
range of merged r1 and r2 if is.overlap(r1,r2) is true, otherwise NA.
}
\description{
Return range of merged intervals if they overlap

This function takes as input two integer vectors. The min and max values of each vector represent the lower and upper limits of each range, respectively. Returns the range of the merged interval if r1 and r2 overrlap, otherwise NA.
}
