% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgrep.R
\name{mgrep}
\alias{mgrep}
\title{Find patterns in subjects}
\usage{
mgrep(query, subject)
}
\arguments{
\item{query}{A character vector of patterns to find}

\item{subject}{A character vector of subjects to search within}
}
\value{
A list numeric vectors. The ith vector indicates which elements of the subject vector contains the ith query.
}
\description{
Find location of each value in a query vector within a subject vector
}
