% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgsub.R
\name{mgsub}
\alias{mgsub}
\title{Replace multiple patterns}
\usage{
mgsub(x, y, z)
}
\arguments{
\item{x}{A character vector of patterns to be replaced}

\item{y}{A character vector of replacement patterns}

\item{z}{Input vector}
}
\value{
output with x values replaced by y values
}
\description{
Replace a set of characters with a set of different characters
}
