% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mround.R
\name{mround}
\alias{mround}
\title{Round in specific way}
\usage{
mround(x, base = 1, direction = "nearest")
}
\arguments{
\item{x}{input number to be rounded}

\item{base}{what number to round by (e.g., base = 5 if "round to the nearest 5")}

\item{direction}{"nearest" (default), alternatively "up" or "down"}
}
\value{
rounded number
}
\description{
Round up or down to the nearest supplied base value
}
