% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.replace.R
\name{na.replace}
\alias{na.replace}
\title{Replace missing values}
\usage{
na.replace(x, replace, ...)
}
\arguments{
\item{x}{Vector possibly containing missing (NA) values.}

\item{replace}{Either a scalar replacement value, or a function returning a scalar value.}

\item{...}{Optional arguments to be passed to replace.}
}
\value{
Vector with missing values (NA) replaced by the value of replace.
}
\description{
Replace missing values. This function is identical to the na.replace function in the gtools package.
Convenience function that is the same as x\link{is.na(x)} <- replace
}
