% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameFromPath.R
\name{nameFromPath}
\alias{nameFromPath}
\title{Filename from Filepath}
\usage{
nameFromPath(path, include.extension = TRUE)
}
\arguments{
\item{path}{Path to file}

\item{include.extension}{If the filename extension should be included in the result (default is TRUE).}
}
\value{
Name of file
}
\description{
Returns the name of a file given the full or relative path to the file, optionally dropping the file extension.
}
