% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmatch.R
\name{nmatch.v2}
\alias{nmatch.v2}
\title{Find First N Matches}
\usage{
nmatch.v2(x, y, n, simplify = F)
}
\arguments{
\item{x}{Vector of values to look for.}

\item{y}{Vector to search within.}

\item{n}{Number of matches to return.}

\item{simplify}{Should the result be coerced to single numeric vector. Default is FALSE.}
}
\value{
when simplify = FALSE, this function returns a list of numeric vectors. The ith vector contains the locations of (at most) the first n matches of x\link{i} in y.
If simplify = TRUE, a single numeric vector is returned that corresponds to concatenating the list of vectors that would hav ebeen returned if simplify=F.
}
\description{
This function returns the indices of the first n matches of each value in a vector x present in a vector y.
This does the nearly the same thing as the function nmatch, but uses base::lapply on base::grep instead of looping base::match
Unlike nmatch, nmatch.v2 has the option of returning a list of vectors holding match indices, rather than a concatenated list of vectors holding match indices.
This method is probably slower than nmatch, but I havent verified this.
}
