% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.inf.sites.R
\name{pairwise.inf.sites}
\alias{pairwise.inf.sites}
\title{Mean Pairwise Informative Sites}
\usage{
pairwise.inf.sites(x, y)
}
\arguments{
\item{x}{input sequence}

\item{y}{second input sequence}
}
\value{

}
\description{
This looks like something Carl wrote. The function is used in the function slice.trim, which might also be from Carl; slice.trim is not used in any of the other functions.
}
