% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.var.aa.R
\name{pars.inf.aa}
\alias{pars.inf.aa}
\title{Test if AA Site Parsimony Informative}
\usage{
pars.inf.aa(alignment.site)
}
\arguments{
\item{alignment.site}{Input character vector representing AA data at one site of an alignment.}
}
\value{
Logical indicating if the site is parsimony informative.
}
\description{
This function is a special case of the function filter.var.aa and is used internally in the function snp.alignment to test if an amino acid alignment site is parsimony informative.
Default values return TRUE for any input character vector.
}
