% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.var.dna.R
\name{pars.inf.dna}
\alias{pars.inf.dna}
\title{Test if DNA Site Parsimony Informative}
\usage{
pars.inf.dna(alignment.site)
}
\arguments{
\item{alignment.site}{Input character vector representing DNA data at one site of an alignment.}
}
\value{
Logical indicating if the site is parsimony informative.
}
\description{
This function is a special case of the function filter.var.dna and is used internally in the function snp.alignment to test if a DNA alignment site is parsimony informative.
Default values return TRUE for any input character vector.
}
