% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pis.new.R
\name{pars.inf}
\alias{pars.inf}
\title{Is Site Parsimony Informative}
\usage{
pars.inf(x)
}
\arguments{
\item{x}{Character vector of DNA at one site of an alignment.}
}
\value{
TRUE if the site is parsimony informative. Otherwise, FALSE.
}
\description{
Given a vector of homologous DNA characters, returns TRUE if the site is parsimony informative and FALSE otherwise.
This is the pars.inf function that was used internally in the pis function of the ips package.
Added "?" to the list of possible characters treated as missing data.
}
