% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent.gaps.R
\name{percent.gaps}
\alias{percent.gaps}
\title{Percentage of Alignment Gaps}
\usage{
percent.gaps(alignment)
}
\arguments{
\item{alignment}{DNA alignment stored as an DNAStringSet}
}
\value{
Percentage of bases that are gaps
}
\description{
Calculates the percentage of bases (not sites) in a DNA alignment that are gaps.
}
