% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent.missing.data.R
\name{percent.missing.data}
\alias{percent.missing.data}
\title{Percent Missing Data}
\usage{
percent.missing.data(alignment)
}
\arguments{
\item{alignment}{DNA alignment stored as an DNAStringSet}
}
\value{
Percent of data missing (gaps or ambiguous) in a DNA sequence alignment
}
\description{
Calculates the percent of bases (not sites) in a DNA alignment that are missing (i.e., gaps or ambiguous).
}
