% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick.loci.R
\name{pick.loci}
\alias{pick.loci}
\title{Pick Rapidly-evolving Exons

This function picks rapidly evolving exons (REEs) to target for sequence capture, given an input table of sequence alignment statistics that was generated using the function makeExomeStatsTable.
Optionally take into account constraints of the bait kit and capture protocol when picking loci, such as bait length, tiling, and number, and/or maximum capture coverage. When using these parameters I suggest overestimating the number of loci that can be captured given the bait kit specifications,
because some of the loci picked with this function will likely be filtered by the company performing bait synthesis (e.g., because baits unstable or difficult to synthesize).}
\usage{
pick.loci(
  statsTable,
  output.path = NULL,
  reference.species = NULL,
  species.subgroup = NULL,
  pident.keep = c(65, 100),
  max.loci.per.gene = NULL,
  min.num.species = "all",
  max.capture.coverage = NULL,
  bait.tiling = NULL,
  bait.length = NULL,
  bait.number = NULL,
  flanking.length = NULL,
  plot.results = FALSE,
  use.min.pident.subgroup = FALSE,
  fast.stat = "pident"
)
}
\arguments{
\item{statsTable}{Data frame or data table object holding the table generated by the makeStatsTable function, or, a character string with the path to this table.}

\item{output.path}{Path where output table should be saved. If NULL (the default), then the output is not written to a file. If the output file already exists a warning message will be generated and the function will terminate.}

\item{reference.species}{The species whose contigs are in the first column of the input table of the statsTable argument. If NULL (the default), "referenceSpecies" is used. For each locus, the other species were aligned to this species when the function makeStatsTable was run.}

\item{species.subgroup}{This is a character vector of species names and should have the same value as what was used for the "subgroup" argument when the function makeStatsTable was run.}

\item{pident.keep}{Minimum and maximum values for min.pident necessary to keep locus.}

\item{max.loci.per.gene}{Maximum number of rapidly-evolving exons (REEs) to target from the same gene. Default is NULL, which means max REEs/gene is unlimited. If set to number, n, then at most only the n fastest-evolving REEs per gene can be included in the output.}

\item{min.num.species}{Either the minumum number of species required to have the locus, or the character string "all" (the default), which means that the locus must be present in all of the species or else it is filtered.}

\item{max.capture.coverage}{Maximum number of nucleotides that can be targeted. Default is NULL.}

\item{bait.tiling}{Fraction overlap between adjacent probes (baits); value of 0.5 corresponds to 50\% overlap between adjacent baits Default is NULL. Ignored if bait.length is NULL.}

\item{bait.length}{Number indicating nucleotide length of baits that will be used for capturing loci. Default is NULL}

\item{bait.number}{Number of unique baits available for sequence capture.This is typically a constraint of the bait kit purchased. Default is NULL.}

\item{flanking.length}{Either NULL (default), "auto", or a number indicating the how many bases upstream and downstream of REEs should also be targeted. Default is NULL.
If set to "auto", flanking lengths are calculated such that the total length of the region targeted is a multiple of the length that would be covered by the number of baits needed to cover the target region, given bait.length and bait.tiling. This argument is ignored if "auto" and either bait.tiling or bait.length are NULL.}

\item{plot.results}{Whether or not results should be plotted. Need to check on what this plot looks like...}

\item{use.min.pident.subgroup}{Whether or not the minimum of the pident.keep argument should be compared to the minimum of pident among the subgroup set of species defined by the species.subgroup argument. Default is FALSE, such that minimum pident.keep is compared to the minimum pident among all species in the input stats table.}

\item{fast.stat}{Character string (or character vector of length 1) indicating which statistic should be used as the metric for "rate of evolution" of loci. The options are:
(1) "pident" (the default) is the mean percent of sites identical to the reference species (which is defined by the reference.species argument); pident is assumed to decrease as evolutionary rate increases.
(2) "percentPIS" is the mean percent of sites that are parsimony informative AND not identical to the reference species; percentPIS is assumed to increase as evolutionary evolutionary rate increases.}
}
\value{
A list holding the input stats table, the input stats table after sorting by fast.stat, and the two stats tables that are also written to files (if write.stats.tables = TRUE). The names of these tables is also included in the output list.
}
\description{
Pick Rapidly-evolving Exons

This function picks rapidly evolving exons (REEs) to target for sequence capture, given an input table of sequence alignment statistics that was generated using the function makeExomeStatsTable.
Optionally take into account constraints of the bait kit and capture protocol when picking loci, such as bait length, tiling, and number, and/or maximum capture coverage. When using these parameters I suggest overestimating the number of loci that can be captured given the bait kit specifications,
because some of the loci picked with this function will likely be filtered by the company performing bait synthesis (e.g., because baits unstable or difficult to synthesize).
}
