% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pis.new.R
\name{pis}
\alias{pis}
\title{Parsimony Informative Sites}
\usage{
pis(x, what = "fraction", use.ambiguities = FALSE)
}
\arguments{
\item{x}{input alignment of DNA sequences stored as either a DNAStringSet, DNAMultipleAlignment, or DNAbin. This is an improved version of the pis function from the phyloch package.}

\item{what}{Either of "absolute", "fraction", or "index", which will return the absolute number, the relative number or the indeces of the potentially-informative sites.}

\item{use.ambiguities}{Should ambiguous bases be considered as non-missing data? Default is false (Don't change this).}
}
\value{
Numeric (depending on what, the number, fraction, or indices of potentially-informative nucleotide sites).
}
\description{
Calculates either the number of parsimony informative sites or the fraction or percent of sites that parsimony inforative.
This is the same function as used in the ips package.
}
