% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAlignment.R
\name{plotAlignment}
\alias{plotAlignment}
\title{Plot Alignment}
\usage{
plotAlignment(alignment)
}
\arguments{
\item{alignment}{Input DNA alignment of class DNAStringSet.}
}
\value{
An object containing the plot that is drawn in the plotting window of R.
}
\description{
Plots the DNA alignment in the plotting window of R. Sequences are shown as lines where data is non-missing.
}
