% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitogenomes.R
\name{plotGBmtc}
\alias{plotGBmtc}
\title{Summarize a GenBank file of mitochondrial genome as a table

Creates a ggplot of the mitochondion that includes features and their locations on heavy and light strands}
\usage{
plotGBmtc(
  pathGB,
  type = "file",
  additionalDF = NULL,
  zoomout = 1.1,
  radii = c(lightStrand = 1, heavyStrand = 0.85),
  linetypes = c(lightStrand = 1, heavyStrand = 1, CDS = 1, stem_loop = 1, misc_feature
    = 1, tRNA = 1, rRNA = 1),
  widths = c(lightStrand = 1, heavyStrand = 2, CDS = 4, stem_loop = 4, misc_feature =
    4, tRNA = 4, rRNA = 4),
  colors = c(lightStrand = "black", heavyStrand = "black", CDS = "green", stem_loop =
    "orange", misc_feature = "yellow", tRNA = "purple", rRNA = "brown"),
  textadj = c(heavy = 0.95, light = 1.04)
)
}
\arguments{
\item{pathGB}{Character string with path to the GenBank flatfile.}
}
\value{
List of length three, which includes [\link{1}] organism name, [\link{2}] ggplot of mitochondrion [\link{3}], data frame used to generate the ggplot
}
\description{
Summarize a GenBank file of mitochondrial genome as a table

Creates a ggplot of the mitochondion that includes features and their locations on heavy and light strands
}
