% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddRADlike.R
\name{proposeLoci.ddRADlike}
\alias{proposeLoci.ddRADlike}
\title{In Silico ddRAD}
\usage{
proposeLoci.ddRADlike(
  input.seqs,
  output.dir,
  recognitionSeqs,
  lim.lengths = c(900, 1000),
  save.tables = T
)
}
\arguments{
\item{input.seqs}{Either NULL (default but not preferred usage), or a character string indicating either a local filepath or URL path to the input sequences in fasta format. The URL character string is often the preferred method.}

\item{output.dir}{Character string indicating the directory where output files should be saved.}

\item{recognitionSeqs}{A character vector of length 2 containing the restriction enzyme recognition sequences to use.}

\item{lim.lengths}{Numerical vector of length 2, containing the minimum and maximum fragment lengths to keep.}

\item{save.tables}{Should the tables holding the locations of each recognition sequence on each contig be saved? Default is TRUE but these tables can be big.}
}
\value{
A table containing the proposed loci and their contig coordinates. Output files (optionally) include two tables, one for each RE recognition sequence, holding the positions of recognition sequences in each genome contig.
}
\description{
Perform in silico ddRAD on a reference genome.
Note: not quite like doing ddRAD because instead of fragmenting and then size selecting, the method implemented here finds ranges on a contig between two different RE cut sites – but these ranges (potential targets) may have cut sites WITHIN them, without consequence.
}
