% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.exome.R
\name{raw.score}
\alias{raw.score}
\title{AA Self-Alignment Score

The alignment score of an amino-acid sequence perfectly aligned to itself under the substitution matrix indicated by "method". Don't trust me on this quite yet!}
\usage{
raw.score(input.seqs, method = "BLOSUM62")
}
\arguments{
\item{input.seqs}{AAString or AAStringSet}

\item{method}{One of the following: "BLOSUM62" (default), "BLOSUM45", "BLOSUM50", "BLOSUM80", "PAM250", "PAM30", or "PAM70".}
}
\value{
Number or numerical vector holding the self-alignment scores for each sequence in input.seqs for the substitution matrix indicated by "method".
}
\description{
AA Self-Alignment Score

The alignment score of an amino-acid sequence perfectly aligned to itself under the substitution matrix indicated by "method". Don't trust me on this quite yet!
}
