% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.gb.R
\name{read.gb}
\alias{read.gb}
\title{Read GenBank flatfile}
\usage{
read.gb(gb.filename, progress = T)
}
\arguments{
\item{gb.filename}{Path to the name GenBank flatfile.}

\item{progress}{Show a progress bar while importing data into R (default is TRUE).}
}
\value{
Object of class gbRecord that contains the information from the GenBank flatfile.
}
\description{
This function imports a GenBank flatfile into R
}
