% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.tree.multipleFiles.R
\name{read.tree.multipleFiles}
\alias{read.tree.multipleFiles}
\title{Read Multiple Newick Trees}
\usage{
read.tree.multipleFiles(filepaths)
}
\arguments{
\item{filepaths}{Vector of filepaths to the newick tree files to read.}
}
\value{
An object of class multiPhylo that holds the trees.
}
\description{
Reads in newick trees from a vector of filepaths.
}
