% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexity.R
\name{repeatScore}
\alias{repeatScore}
\title{Repeat Score

This function calculates fraction of bases in a DNA sequence that initiate a specified pattern. At present only the default pattern (base\link{i} == base\link{i+2}) can be used.}
\usage{
repeatScore(x, rp = "1.1")
}
\arguments{
\item{x}{Character string, or a DNAString object, or a length 1 DNAStringSet object as defined by the Biostrings package.}

\item{rp}{Text string defining that defines how complexity is defined. \link{AT PRESENT ONLY PATTERNS COMPOSED OF "1" and "." CAN BE USED}. Default is "1.1", which specifies to return the fraction of bases in which base\link{i} == base\link{i+2}.}
}
\value{
Number between 0 and 1 indicating the repeat score for the repeat pattern indicated by rp. A value of 1 indicates that all possible reading frames and frame position having a width equal to the rp pattern width have the rp pattern. A value of zero indicates that the pattern is absent.
}
\description{
Repeat Score

This function calculates fraction of bases in a DNA sequence that initiate a specified pattern. At present only the default pattern (base\link{i} == base\link{i+2}) can be used.
}
