% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportBestMatches.R
\name{reportBestMatches}
\alias{reportBestMatches}
\title{Get Sequences From BLAST Table}
\usage{
reportBestMatches(
  input.table,
  output.table.path = NULL,
  remove.subseq.matches = F,
  min.bitscore = 50,
  min.bitscore.difference = 0,
  blast.method = NULL,
  table.format = 6
)
}
\arguments{
\item{input.table}{Character string with path to BLAST hit table, or, a BLAST hit table held as an object of class data.table, data.frame, or matrix.}

\item{output.table.path}{Character to string with path where output table should be saved; default is NULL.}

\item{remove.subseq.matches}{Logical with whether or not to drop matches for which the subject sequence is a subsequence of another match. Default is FALSE.}

\item{min.bitscore}{Number indicating the minimum bitscore required to keep the best match. Matches with a bitscore below this value are not included in the output table. This is useful for removing paralog matches. Default = 50.
The reason for using a default min.bitscore of 50 is because values at or greater than this usually indicate that the sequences are homologous (doi: 10.1002/0471250953.bi0301s42).}

\item{min.bitscore.difference}{Number indicating the minimum difference required between bitscores of the best and second best matches; must be greater than this value to keep matches for the query sequence. Default = 0. This is useful for removing loci with putatitive recent duplicates in the genome.}

\item{blast.method}{At present, this argument is ignored unless set to "tblastn", in which case the translation frame suffix is removed from query sequence IDs (if present). In the future, I will either delete this argument or this argument will take as input one of the following: NULL (default), or one of "blastn", "blastp", "blastx", "tblastn", "tblastx".}

\item{table.format}{Number indicating which NCBI table format is applicable to input.table; At present, only format 6 can be used. If more than 12 columns are present only the first 12 are used and assumed to correspond to the columns of format 6.}
}
\value{
A data.table object containing the best matche to each input query in the input blast table. If output.table.path argument is a path (character string), then the function also writes the output as a tab-separated file.
}
\description{
Filter a blast hit table of matches to include only the best matches (highest bitscore) to each query.
Need to add ability to:
(2) Drop query sequences if the best match has a bitscore below some threshold.
}
