% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRequired.R
\name{required.packages}
\alias{required.packages}
\title{State Required Packages}
\usage{
required.packages(packages, include.input = F, include.Rcore = T)
}
\arguments{
\item{packages}{Character vector of package names}

\item{include.input}{Should the output vector also include the input packages? Default FALSE.}

\item{include.Rcore}{Should the output vector include R-core packages? Default TRUE.}
}
\value{
Character vector of names of packages required (directly or indirectly) by the input set of packages.
}
\description{
Get a list of all packages required (imports and depends) for a vector of packages.
Package versions are ignored.
}
