% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollSum.R
\name{rollSum}
\alias{rollSum}
\title{Rolling Summation}
\usage{
rollSum(x)
}
\arguments{
\item{x}{Input numerical vector}
}
\value{
Output is a numerical vector that is the rolling summation of the input vector x
}
\description{
Returns the rolling summation vector of input numerical vector
}
