% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.trimal.R
\name{run.trimal}
\alias{run.trimal}
\title{Run trimal}
\usage{
run.trimal(
  input.align,
  method = "auto",
  locus.name = locus.names[i],
  trimal.exe.dir,
  trimal.exe.name = "trimal"
)
}
\arguments{
\item{input.align}{Input DNA alignment (class DNAStringSet)}

\item{method}{Method to use (default is "auto"). Currently "auto" is the only method implemented in this function, and runs trimal with the "-automated1" flag to "Use a heuristic selection of the automatic method based on similarity statistics." - (http://trimal.cgenomics.org/use_of_the_command_line_trimal_v1.2)}

\item{locus.name}{Name of the locus held in the input alignment (default is locus.names\link{i}). If this is kept at the default value then the function should be run within a loop and there needs to be a vector called locus names.}

\item{trimal.exe.dir}{Directory where the trimal executable is located.}

\item{trimal.exe.name}{Name of the trimal executable (default is "trimal")}
}
\value{
trimAL alignment (class DNAStringSet)
}
\description{
Wrapper for running trimal.
"trimAl is a tool for the automated removal of spurious sequences or poorly aligned regions from a multiple sequence alignment." - (http://trimal.cgenomics.org/).
This function takes an input DNA alignment of class DNAStringSet and saves a temporary copy in fasta format. Then, the function runs trimAL on the fasta formatted alignment and saves a temporary trimAL alignment in the current directory.
Next, reads the trimAL alignment into R and deletes it from the current directory.
Value returned by this function is the trimAL alignment stored as an object of class DNAStringSet.
Original version Code written by Carl Hutter?
}
