% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanFa.multiple.R
\name{scanFa.multiple}
\alias{scanFa.multiple}
\title{Read Multiple Fasta Files}
\usage{
scanFa.multiple(directory)
}
\arguments{
\item{directory}{Path to folder containing multiple fasta formatted sequence files.}
}
\value{
A DNAStringSet, RNAStringSet, or AAStringSet (determined by the sequences).
}
\description{
Read fasta formatted sequences stored in multiple files in the same directory.
}
