% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatenateDNAStringSets.R
\name{simple.concatenate.DNA}
\alias{simple.concatenate.DNA}
\title{Simple Concatenate DNA Alignment}
\usage{
simple.concatenate.DNA(..., make.partitions.table = F)
}
\arguments{
\item{...}{Two or more DNAStringSet objects separated by commas. All taxa must be present in all input DNAStringSets. All sequences in each DNAStringSet must have the same length and are assumed to represent homologous sites.}

\item{make.partitions.table}{Logical indicating if a partition table should also be created. Default is FALSE. This argument is not yet implemented.}
}
\value{
A concatenated DNA alignment held in DNAStringSet object.
}
\description{
Used to concatenate multiple DNAStringSet or DNAMultipleAlignment files. Concatenation order is the same as the input argument order.
Optionally creates a partition table (option not yet implemented).
Compare with function concatenateDNAStringSets.
}
