% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.trim.R
\name{slice.trim}
\alias{slice.trim}
\title{Remove Poorly Aligned Regions}
\usage{
\method{slice}{trim}(input.align, slice.size.bp = 100, threshold = 0.45, min.length = 20)
}
\arguments{
\item{input.align}{Input DNA alignment of class DNAStringSet or DNAMultipleAlignment}

\item{slice.size.bp}{Window size for the sliding window. Default is 100.}

\item{threshold}{Maximum fraction of sites that can be parsimony informative sites (mean of pairwise scores calculated for each individual) for the sequence to be considered aligned "good".}

\item{min.length}{Minimum length non-missing data to keep an individual in the output alignment.}
}
\value{
DNA alignment (class DNAStringSet) with poorly aligned regions replaced with missing data.
}
\description{
A sliding window method to replace poorly aligned subsections of an alignment (these are replaced with a string of "-"). Additionally, individuals with less non-missing data than a threshold are removed.
Did I or Carl Hutter write this code?
}
