% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp.alignment.R
\name{snp.alignment}
\alias{snp.alignment}
\title{Get SNP Alignment}
\usage{
snp.alignment(alignment)
}
\arguments{
\item{alignment}{Input alignment}
}
\value{
An alignment containing only the SNPs of the input alignment.
}
\description{
Filters an input alignment and returns an alignment containing only the sites that are SNPs.
This function is the same as running filter.alignment(alignment,min.allele.freqs.dna=c(2,2,0,0))
}
