% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp.alignment.R
\name{snpsDNAStringSet}
\alias{snpsDNAStringSet}
\title{SNPs DNAStringSet Alignment}
\usage{
snpsDNAStringSet(alignment, remove.NoDataSeqs = T)
}
\description{
Removes non-SNP columns of a DNA alignment of class DNAStringSet. Does the same thing as snp.alignment function, but this function is more efficient although only works on DNAStringSet objects.
}
\details{
@param alignment DNA sequence alignment of class DNAStringSet
@param remove.NoDataSeqs Filter the SNPs alignment to only include individuals with at least one A, C, G, or T.
@return DNAStringSet alignment of SNPs
@export
}
