% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_locate_last.R
\name{str_locate_last}
\alias{str_locate_last}
\title{Locate Last in Strings}
\usage{
str_locate_last(strings, pattern)
}
\arguments{
\item{strings}{One or more character strings to search within}

\item{pattern}{Character string to search for}
}
\value{
Integer vector indicating position of last occurrence of pattern in strings.
}
\description{
Character position of the last occurrence of pattern if the pattern is >1 character long, the function returns the position of the last character of the last instance of the pattern
}
