% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_locate_X.R
\name{str_locate_X}
\alias{str_locate_X}
\title{Locate Xth Occurrence in String}
\usage{
str_locate_X(strings, pattern, X)
}
\arguments{
\item{strings}{One or more character strings to search within}

\item{pattern}{Character string to search for}

\item{X}{Integer indicating which occurrence to find}
}
\value{
Integer vector indicating position of last occurrence of pattern in strings.
}
\description{
Character position of the Xth occurrence of a pattern. If the pattern is >1 character long, the function returns the position of the last character of the Xth occurrence.
}
