% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.list.R
\name{string.list}
\alias{string.list}
\title{Matrix to List of Strings}
\usage{
string.list(y)
}
\arguments{
\item{y}{Input character matrix}
}
\value{
Returns a list character strings, each string formed by the row elements of the input matrix
}
\description{
Collapses rows of a character matrix into character strings
}
