% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.partitioned.alignment.R
\name{submat}
\alias{submat}
\title{submat

Return a substitution matrix for AA or DNA sequences to be used with the Biostrings::stringDist function.
Ambiguous nucleotides or amino acids are treated as missing data and are ignored.}
\usage{
submat(type = "DNA")
}
\arguments{
\item{type}{One of: "DNA","AA", "DNA_comparable", "AA_comparable"}
}
\value{
XStringSet object with same class as object supplied to 'xstrings', with leading and trailing polyNs/polyXs removed or replaced depending on the value supplied to 'repl'
}
\description{
submat

Return a substitution matrix for AA or DNA sequences to be used with the Biostrings::stringDist function.
Ambiguous nucleotides or amino acids are treated as missing data and are ignored.
}
\examples{

Condsider the alignment:
> alignment
    width  seq                       names
[1]    25  -----ACGTACGTAC-ACTG-----  Seq1
[2]    25  ACGTAACGTACGTAC-AC-------  Seq2
[3]    25  CGTACA--TACGTAC-ACTGACTGN  Seq3

# Trimming alignment to Seq1:
> trimTo(aln=alignment,c("Seq1"))
    width  seq             names
[1]    15  ACGTACGTAC-ACTG  Seq1
[2]    15  ACGTACGTAC-AC--  Seq2
[3]    15  A--TACGTAC-ACTG  Seq3

# Trimming alignment to Seq2:
> trimTo(aln=alignment,c("Seq2"))
    width  seq                names
[1]    18  -----ACGTACGTAC-AC  Seq1
[2]    18  ACGTAACGTACGTAC-AC  Seq2
[3]    18  CGTACA--TACGTAC-AC  Seq3

# Trimming alignment to Seq1 and Seq2:
> trimTo(aln=alignment,c("Seq1","Seq2"))
    width  seq                  names
[1]    20  -----ACGTACGTAC-ACTG  Seq1
[2]    20  ACGTAACGTACGTAC-AC--  Seq2
[3]    20  CGTACA--TACGTAC-ACTG  Seq3
}
