% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitogenomes.R
\name{summarizeGBmtc}
\alias{summarizeGBmtc}
\title{Summarize a GenBank file of mitochondrial genome as a table

Creates a table summizing the annotations held in the GenBank flatefile. This function is used in the function plotGBmtc.}
\usage{
summarizeGBmtc(pathGB)
}
\arguments{
\item{pathGB}{Character string with path to the GenBank flatfile.}
}
\value{
List of length five, which includes [\link{1}] organism name, [\link{2}], organelle name [\link{3}], length of the sequence [\link{4}] heavy and light strand labels, and [\link{5}] a table with features and their ranges.
}
\description{
Summarize a GenBank file of mitochondrial genome as a table

Creates a table summizing the annotations held in the GenBank flatefile. This function is used in the function plotGBmtc.
}
