% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_col.R
\name{t_col}
\alias{t_col}
\title{Make Transparent Color}
\usage{
t_col(color, percent = 50, name = NULL)
}
\arguments{
\item{color}{color name}

\item{percent}{\% transparency (default is 50)}

\item{name}{Optional name for the color (default is NULL)}
}
\value{
Returns transparent version of a named color.
}
\description{
Makes a transparent version of a named color. Function is from "Transparent colors. Mark Gardener 2015. www.dataanalytics.org.uk"
}
