% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PostsequencingPipeline.R
\name{targetMatchingAssessment03b}
\alias{targetMatchingAssessment03b}
\title{Step 3b: target matching assessment}
\usage{
targetMatchingAssessment03b(
  workdir = getwd(),
  targets.path,
  minOverlap = 41,
  maxeval = 0.05,
  minbitscore = 50
)
}
\arguments{
\item{workdir}{Path to the directory containing the 'Processed_Samples' folder used in steps 1–3 of SnakeCap. Default is the current directory.}

\item{targets.path}{Path to the file containing the loci that were targetted for capture, and from which probes were designed.}

\item{minOverlap}{Minimum number of matching bases required between the contig and reference target to retain the match. Default 41.}

\item{maxeval}{Maximum e-value allowed to retain a match. Default 0.05.}

\item{minbitscore}{Minimum bitscore allowed to retain a match. Default 50.}
}
\description{
This function performs step 3b of the post-sequencing pipeline used for SnakeCap
In this step, haplocontig consensus sequences generating during step 2 (dipspades.sh), target sequences used to generate probes prior to sequencing, and blast hit tables generated during step 3a (targetMatching.sh) are processed to sort/filer assembled consensus haplocontigs
}
