% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.exome.R
\name{translate.exome}
\alias{translate.exome}
\title{translate CDS regions

Used to translate the output of get.seqs.from.gff. Translates reading frame 1 of forward and reverse complement sequences of input DNAStringSet and returns whichever has fewer stop codons.
The output of this function can be the query input for blast function with method "tblastn"}
\usage{
translate.exome(input.seqs, drop.if.stop = T)
}
\arguments{
\item{input.seqs}{DNAStringSet produced by get.seqs.from.gff}

\item{drop.if.stop}{Should AA sequences with internal stop codons be dropped? Default is TRUE.}
}
\value{
AAStringSet containing translated sequences of input.seqs
}
\description{
translate CDS regions

Used to translate the output of get.seqs.from.gff. Translates reading frame 1 of forward and reverse complement sequences of input DNAStringSet and returns whichever has fewer stop codons.
The output of this function can be the query input for blast function with method "tblastn"
}
