% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.ends.R
\name{trim.ends}
\alias{trim.ends}
\title{Trim ends of DNA alignment}
\usage{
trim.ends(x, min.n.seq = 4, codon.trim = T)
}
\arguments{
\item{x}{DNA alignment held in an object of class DNAStringSet.}

\item{min.n.seq}{Minimum number of individuals with non-missing data at site required to keep the site (default is 4).}

\item{codon.trim}{Trim alignment so that the first site is codon 1 and last site is codon 3. (default is TRUE).}
}
\value{
Trimmed alignment of class DNAStringSet
}
\description{
This function is PROBABLY used to create an alignment for input into a function that translates CDS DNA.
For other (most) applications, the functions "trim.alignment" or "filter.alignment" should be used instead.
Need to look further into when/if this function is used in my REEs pipeline.
Not sure if I wrote this or if this is Carl Hutter code.
}
