% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.partitioned.alignment.R
\name{trimTo}
\alias{trimTo}
\title{Trim alignment to covered by spanned by indidividual

Trims an alignment to the largest window containing all non-gap characters of a subset of individuals. See examples.}
\usage{
trimTo(aln, nam)
}
\arguments{
\item{aln}{Object of class XStringSet (Biostrings package; DNAStringSet, AAStringSet, or RNAStringSet)}

\item{nam}{Character string vector with one or more names of sequences in aln}
}
\value{
DNAStringSet
}
\description{
Trim alignment to covered by spanned by indidividual

Trims an alignment to the largest window containing all non-gap characters of a subset of individuals. See examples.
}
