% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.partitioned.alignment.R
\name{trimXN}
\alias{trimXN}
\title{trimXN

Remove or replace 5' and 3' polyNs and polyXs}
\usage{
trimXN(xstrings, repl = "")
}
\arguments{
\item{xstrings}{Object of class XStringSet (Biostrings package; DNAStringSet, AAStringSet, or RNAStringSet)}

\item{repl}{Either an empty string (default; polyNs and polyXs), or a single character that will replace each N or X of 5' and 3' polyNs and polyXs}
}
\value{
XStringSet object with same class as object supplied to 'xstrings', with leading and trailing polyNs/polyXs removed or replaced depending on the value supplied to 'repl'
}
\description{
trimXN

Remove or replace 5' and 3' polyNs and polyXs
}
\examples{
input.alignments.directory <- "~/Immune/unpartitioned/"
output.directory           <- "~/Immune/partitioned/"
target.loci                <- "~/Weinell_TargetLoci_Snakes_Final_targetCDS_v3.fa"
input.baits.table          <- "~/bait_species_table.txt"

make.partitioned.alignment(InputAlignmentFolder=input.alignments.directory, output.dir=output.directory, TargetCDS.path=target.loci, bait.species.filename=input.baits.table)
}
