% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique.tiplabels.R
\name{unique.tiplabels}
\alias{unique.tiplabels}
\title{All Unique Tree Tips}
\usage{
\method{unique}{tiplabels}(trees.multiPhylo)
}
\arguments{
\item{trees.multiPhylo}{Input set of trees held in an object of class multiPhylo (see ape Package for more details on multiPhylo class objects)}
}
\value{
Character vector of tree tip labels occurring in at least one tree in the
}
\description{
Return the unique set of tip labels for an input set of trees
}
