% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexity.R
\name{wapply}
\alias{wapply}
\title{Apply function as sliding window

This function applies a function to a DNA sequence as a sliding window.}
\usage{
wapply(
  x,
  FUN,
  windowsize = NULL,
  windowshift = NULL,
  startbase = 1,
  remainder = T,
  minwindowsize = 2
)
}
\arguments{
\item{x}{Character string, or a DNAString object, or a length 1 DNAStringSet object as defined by the Biostrings package.}

\item{FUN}{The function to apply to the DNA sequence framed by each position of the window.}

\item{windowsize}{The size of the window in base pairs. Default value NULL, in which case window size is equal to the length of x.}

\item{windowshift}{The distance to move the window between steps. Default value NULL, in which case windowshift=windowsize.}

\item{startbase}{Number indicating the first base to include in the first frame position. Default 1.}

\item{remainder}{Logical indicating whether to apply the function to the last window if the window size is less than indicated by the windowsize argument. Default TRUE.}

\item{minwindowsize}{Number indicating the minimum allowable window size. Ignored if windowsize < minwindowsize.}
}
\value{
A list or vector, with the ith entry equal to the value returned by applying FUN to the ith window of the input DNA sequence.
}
\description{
Apply function as sliding window

This function applies a function to a DNA sequence as a sliding window.
}
