% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.phy.R
\name{write.phy}
\alias{write.phy}
\title{Write DNA phylip format}
\usage{
write.phy(x, file = "", interleave = FALSE, strict = FALSE)
}
\arguments{
\item{x}{input sequences. Each sequence is a character string?}

\item{file}{Where to write sequences}

\item{interleave}{Whether or not sequences should be written interleaved (default is false)}

\item{strict}{Whether or not taxanames should be limited to 10 characters (default is false)}
}
\value{

}
\description{
Write DNA sequences in phylip format. This is EXACTLY the write.phy function in the package phyloch.
}
