% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeDNAStringSet.R
\name{writeDNAStringSet}
\alias{writeDNAStringSet}
\title{Write DNAStringSet}
\usage{
writeDNAStringSet(
  x,
  filepath,
  append = F,
  format = "fasta",
  charsPerLine = 1e+05
)
}
\arguments{
\item{x}{Input DNAStringSet object (see Biostrings package for details on this class)}

\item{filepath}{Where to save sequences}

\item{append}{Should the sequences be appended to a file that already exists? (default is FALSE)}

\item{format}{Format to save sequences. Available formats are "fasta" (default) or "phylip".}

\item{charsPerLine}{Maximum number of bp to save per line (default is 100,000 and intended for writing sequences sequentiall rather than interleaved). Only invoked when format is fasta.}
}
\value{
Writes the input DNAStringSet to filepath. Returns the input DNAStringSet x.
}
\description{
Write a set of DNA sequences to file in sequential fasta or phylip format.
}
