% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xprod.combn.mat.R
\name{xprod.combn.mat}
\alias{xprod.combn.mat}
\title{All Combinations as Matrix}
\usage{
xprod.combn.mat(..., unique.mode = "index")
}
\arguments{
\item{...}{Two or more character vectors (or any vector type coercible to character class). Vectors can be any non-zero length. If input vectors are very long this function may crash. Similarly (and moreso), a large number of input vectors will likely crash this function.}

\item{unique.mode}{One of types "unordered", "ordered", or "index" (default is "index").
If unique.mode="unordered", output sets (rows of output matrix) are considered equivalent if they contain the same elements regardless of order, i.e. set (A,B) = set (B,A), and only the former will be returned.
If unique.mode="ordered", output sets are considered equivalent if they contain the same elements in the same order.
If unique.mode="index". Output sets are retained even if they contain the same elements in the same order, which occurs when at least one input vector contains a repeated value.}
}
\value{
Character matrix with as many columns as input vectors. Each row is a unique combination (as defined by unique.mode parameter) of elements from the input vectors.
}
\description{
Find all n-wise combinations of elements from n input sets.
}
