% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xprod.combn.R
\name{xprod.combn}
\alias{xprod.combn}
\title{All Combinations as Strings}
\usage{
xprod.combn(..., unique.mode = "index")
}
\arguments{
\item{...}{Two or more character vectors (or any vector type coercible to character class). Vectors can be any non-zero length. If input vectors are very long this function may crash. Similarly (and moreso), a large number of input vectors will likely crash this function.}

\item{unique.mode}{One of types "unordered", "ordered", or "index" (default is "index").
If unique.mode="unordered", output sets (character strings with elements separate by " ") are considered equivalent if they contain the same elements regardless of order, i.e. set (A,B) = set (B,A), and only the former will be returned.
If unique.mode="ordered", output sets are considered equivalent if they contain the same elements in the same order.
If unique.mode="index". Output sets are retained even if they contain the same elements in the same order, which occurs when at least one input vector contains a repeated value.}
}
\value{
A character vector. Each vector element is a character string of the input vector elements separated by a space " ". Definition of unique combination is controlled by the unique.mode parameter.
}
\description{
Find all n-wise combinations of elements from n input sets.
}
