% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.seqs.from.blastTable.R
\name{get.seqs.from.blastTable}
\alias{get.seqs.from.blastTable}
\title{Get Sequences From BLAST Table

Extracts the sequences defined in an input table (a blast table with NCBI format 6) from an input genome. The input genome should be the same genome used as the subject in the blast table.
The quality/completeness of the output sequence dataset depends on the quality/completeness of the blast search.
This function supercedes the functions get.exome.from.blastTable and get.UCEs.from.blastTable.}
\usage{
get.seqs.from.blastTable(input.blastTable, input.seqs, output.path = NULL)
}
\arguments{
\item{input.blastTable}{Either a character string with the local filepath to a BLAST table (BLAST output format 6), or such a table held as a data table, data frame, or character matrix object.}

\item{input.seqs}{Either a character string with the filepath or URL path to the input sequences (which can be .gz compressed), or, a DNAStringSet object holding the sequences.}

\item{output.path}{Character string with path (including filename) to where extracted sequences should be saved.}
}
\value{
A DNAStringSet object with the extracted sequences. Sequences are also optionally written to output.path.
}
\description{
Get Sequences From BLAST Table

Extracts the sequences defined in an input table (a blast table with NCBI format 6) from an input genome. The input genome should be the same genome used as the subject in the blast table.
The quality/completeness of the output sequence dataset depends on the quality/completeness of the blast search.
This function supercedes the functions get.exome.from.blastTable and get.UCEs.from.blastTable.
}
