% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.seqs.from.gff.R
\name{get.seqs.from.gff}
\alias{get.seqs.from.gff}
\title{Get Sequences from GFF Table}
\usage{
get.seqs.from.gff(
  input.seqs,
  input.gff,
  output.path = NULL,
  additional.ID = NULL
)
}
\arguments{
\item{input.seqs}{Character string indicating the local or URL path to input sequences.}

\item{input.gff}{Either a data table object, or character string  or local path to the filtered gff containing the the loci of interest.}

\item{output.path}{Path where exomes should be saved (default is NULL).}

\item{additional.ID}{Filename of a table that cross-references contig names to names in GFF file (default is NULL).}
}
\value{
DNAStringSet with sequences corresponding to unique ranges in input.gff; sequences are written to output.file in fasta format if output.path is not NULL.
}
\description{
Identical to the function get.exome.from.gff, but used more generally.
Note to self: I used this function to extract the MHC loci from the Thamnophis sirtalis genome.
}
