% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBlastDB.R
\name{makeBlastDB}
\alias{makeBlastDB}
\title{Make Local Blast Database

Wrapper for the NCBI makeblastdb command. This function makes an NCBI blast database (which is needed to run blast against).}
\usage{
makeBlastDB(makeblastdb.path, subject.path)
}
\arguments{
\item{makeblastdb.path}{Path to makeblastdb (included in the bin folder of blast)}

\item{subject.path}{Path to the file containing the sequences to include in the database}
}
\value{
Creates a blast database file.
}
\description{
Make Local Blast Database

Wrapper for the NCBI makeblastdb command. This function makes an NCBI blast database (which is needed to run blast against).
}
