% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeStatsTable.R
\name{makeStatsTable}
\alias{makeStatsTable}
\title{Make Stats Table

This function aligns and then creates a table of stats calculated for each exon alignment. The output stats table is used to identify/select the REEs.
This function should be used instead of running the functions align.and.concatenate.best.exons and then alignment.stats}
\usage{
makeStatsTable(
  input.seqs,
  species.names,
  reference.species,
  input.gff,
  output.path = NULL,
  alignments.out = NULL,
  subgroup = NULL,
  i.start = 1,
  i.stop = NA,
 
    mafft.params = "--localpair --maxiterate 1000 --adjustdirection --quiet --op 3 --ep 0.123 --thread 6"
)
}
\arguments{
\item{input.seqs}{Either a list of DNAStringSet objects with each species sequences, or a vector of character strings with paths to input sequences in fasta format.}

\item{species.names}{Vector of character strings containing names of species associated with each input.seqs.}

\item{reference.species}{Number indicating which species is the reference species.}

\item{input.gff}{One of the following: An object of class data table, data.frame, or character matrix, with columns identical to those used in Generic Feature Format (GFF) tables; or, a character string with local file path to either a GFF file or table with columns identical to those used in GFF files; or, a character string with URL path to a GFF file; or, NULL, in which case the gene.names column of the output table will be filled with NAs.
This table can either be the original GFF table or the one generated by the function filter.gff. Exon and protein isoform information may included in future versions of the output table, and this would require the original GFF table.
If not NULL, the GFF must correspond to the reference species.}

\item{output.path}{Where to save the stats table. If NULL, the stats table is saved to a temporary file.}

\item{alignments.out}{Directory where to save output alignment. Default is NULL (alignments not saved). If saved, alignments are in fasta format.}

\item{subgroup}{Optional vector of character strings containing a subset of the "species" argument. Default is NULL. This parameter determines which species are used for calculating stats. If NULL, all species in the species argument are used.}

\item{i.start}{First locus to start at. Default is 1.}

\item{i.stop}{Last locus to include. Default is NA, meaning to include all loci from i.start.}

\item{mafft.params}{Character string of mafft parameter settings. Default is "--localpair --maxiterate 1000 --adjustdirection --quiet --op 3 --ep 0.123 --thread 6". I don't suggest changing these.}
}
\value{
Stats table (class data.table object), which is also saved to the value of output.path.
}
\description{
Make Stats Table

This function aligns and then creates a table of stats calculated for each exon alignment. The output stats table is used to identify/select the REEs.
This function should be used instead of running the functions align.and.concatenate.best.exons and then alignment.stats
}
