% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick.loci.R
\name{pick.loci}
\alias{pick.loci}
\title{Pick Rapidly-evolving Exons}
\usage{
pick.loci(
  statsTable.path,
  primary.species,
  output.path = NULL,
  species.subgroup = NULL,
  pident.keep = c(65, 100),
  max.loci.per.gene = NULL,
  min.num.species = "all",
  max.capture.coverage = 1200000,
  plot.results = FALSE,
  use.min.pident.subgroup = FALSE,
  fast.stat = "pident"
)
}
\arguments{
\item{statsTable.path}{Path to the input stats table, which is the table generated by the function makeExomeStatsTable. Example: "stats_exome_data_TBLASTX.txt".}

\item{primary.species}{The species whose contigs are in the first column of the input table indicated by the statsTable.path argument. For each locus, the other species were aligned to this species when the function makeStatsTable was run.}

\item{output.path}{Path where output table should be saved. If NULL (the default), then the output is not written to a file. If the output file already exists a warning message will be generated and the function will terminate.}

\item{species.subgroup}{This is a character vector of species names and should have the same value as what was used for the "subgroup" argument when the function makeStatsTable was run.}

\item{pident.keep}{Minimum and maximum values for min.pident necessary to keep locus.}

\item{max.loci.per.gene}{Maximum number of rapidly-evolving exons (REEs) to target from the same gene. Default is NULL, which means max REEs/gene is unlimited. If set to number, n, then at most only the n fastest-evolving REEs per gene can be included in the output.}

\item{min.num.species}{Either a number that indicates the minimum number of species with the locus, or else the locus is filtered, or the character string "all" (the default), which means that the locus must be present in all of the species.}

\item{max.capture.coverage}{Maximum number of nucleotides that can be targeted. Default is 1200000.}

\item{plot.results}{Whether or not results should be plotted. Need to check on what this plot looks like...}

\item{use.min.pident.subgroup}{Whether or not the minimum of the pident.keep argument should be compared to the minimum of pident among the subgroup set of species defined by the species.subgroup argument. Default is FALSE, such that minimum pident.keep is compared to the minimum pident among all species in the input stats table.}

\item{fast.stat}{Character string (or character vector of length 1) indicating which statistic should be used as the metric for "rate of evolution" of loci. The options are:
(1) "pident" (the default) is the mean percent of sites identical to the primary species (which is defined by the primary.species argument); pident is assumed to decrease as evolutionary rate increases.
(2) "percentPIS" is the mean percent of sites that are parsimony informative AND not identical to the primary species; percentPIS is assumed to increase as evolutionary evolutionary rate increases.}
}
\value{
A list holding the input stats table, the input stats table after sorting by fast.stat, and the two stats tables that are also written to files (if write.stats.tables = TRUE). The names of these tables is also included in the output list.
}
\description{
This function picks rapidly evolving exons (REEs) to target for sequence capture, given an input table of sequence alignment statistics that was generated using the function makeExomeStatsTable.
}
