% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportBestMatches.R
\name{reportBestMatches}
\alias{reportBestMatches}
\title{Get Sequences From BLAST Table}
\usage{
reportBestMatches(
  input.table,
  output.table.path = NULL,
  remove.subseq.matches = T,
  min.bitscore = 50,
  min.bitscore.difference = 0
)
}
\arguments{
\item{input.table}{Path to the input hit table of matches, or, a data.table object.}

\item{output.table.path}{Where to save the output table of best matches (default is NULL).}

\item{remove.subseq.matches}{Whether or not to drop matches for which the subject sequence is a subsequence of another match. Default is TRUE.}

\item{min.bitscore}{Minimum bitscore required for the best match. Matches with a bitscore below this value are not included in the output table. This is useful for removing paralog matches. Default = 50.
The reason for using a default min.bitscore of 50 is because values at or greater than this usually indicate that the sequences are homologous (doi: 10.1002/0471250953.bi0301s42).}

\item{min.bitscore.difference}{The difference between the best and second best matches' bitscore must be greater than this value to keep matches for the query sequence. Default = 0. This is useful for removing loci with putatitive recent duplicates in the genome.}
}
\value{
A data.table object containing the best matche to each input query in the input blast table. If output.table.path argument is a path (character string), then the function also writes the output as a tab-separated file.
}
\description{
Filter a blast hit table of matches to include only the best matches (highest bitscore) to each query.
Need to add ability to:
(2) Drop query sequences if the best match has a bitscore below some threshold.
}
